% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_look.R
\name{add_ci}
\alias{add_ci}
\title{Add a confidence interval look to a shape.}
\usage{
add_ci(
  pj,
  m,
  type = "line",
  dir = "v",
  d = NA,
  pos = NA,
  neg = NA,
  shape = ".",
  name = "."
)
}
\arguments{
\item{pj}{an object of class \code{'pj'}.}

\item{m}{positive numeric vector of length recyclable with \code{x} and
\code{y} defining the multiplier to apply to confidence interval distances
in \code{d} or \code{pos} and \code{neg}.}

\item{type}{\code{'line'}, \code{'cap'}, \code{'taper'}, or \code{'ribbon'}.
See the \emph{types of confidence intervals} section.}

\item{dir}{\code{'h'} for horizontal intervals or \code{'v'} for vertical
intervals.}

\item{d}{\code{NA} or a positive numeric vector of length recyclable with
\code{x} and \code{y} defining the symmetric positive/negative error
magnitude at each midpoint. See \code{the d, pos, and neg arguments}
section.}

\item{pos}{\code{NA} or positive numeric vector of length recyclable with
\code{x} and \code{y} defining positive error magnitude at each midpoint.
See \code{the d, pos, and neg arguments} section.}

\item{neg}{\code{NA} or positive numeric vectors of length recyclable with
\code{x} and \code{y} defining negative error magnitude at each midpoint.
See \code{the d, pos, and neg arguments} section.}

\item{shape}{character scalar indicating the type of shape to which the
look will be applied.}
}
\value{
\code{pj} with the addition of the confidence interval look applied
to the specified points shape.
}
\description{
Confidence interval looks can only be added to point shapes.
}
\section{Types of confidence intervals}{

\cr\emph{line}
\cr A line confidence interval is simply a line between the minimum point
of a confidence interval and its maximum point.
\cr\cr
\emph{cap}
\cr A capped-line confidence interval is a line interval with a cap at the
ends of the line.
\cr\cr
\emph{taper}
\cr A tapered confidence interval is a thin polygon thickest at the middle
of the confidence interval and narrowing to zero width at the lower and
upper bounds of the interval.
\cr\cr
\emph{ribbon}
\cr A ribbon confidence interval is a polygon defined by the area between
minimum and maximum values of confidence intervals for a sorted series of
points.
}

\section{The d, pos, and neg arguments}{
 Either \code{d} must be supplied or
both \code{pos} and \code{neg} must be supplied.
}

\section{Valid combinations of argument properties}{

Intervals can be validly described using a number of patterns of argument
properties as follows: \describe{
\item{\code{x + y + m + dir + type + d}}{symmetric confidence interval of
width \code{2dm}}.
\item{\code{x + y + m + dir + type + neg + pos}}{non-symmetric confidence
interval of width \code{m(pos + neg)}}.
}
}

\section{Argument recycling}{
 The argument \code{d} or the arguments
\code{pos} and \code{neg} are recycled with the values of \code{x} and
\code{y} associated with the points shape the confidence interval is to be
applied to.
}

\seealso{
Other add_look: 
\code{\link{add_ends}()},
\code{\link{add_fills}()},
\code{\link{add_glyphs}()},
\code{\link{add_labels}()},
\code{\link{add_lines}()},
\code{\link{add_marks}()}
}
\concept{add_look}
