% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_look.R
\name{add_ends}
\alias{add_ends}
\title{Add an ends look to a shape.}
\usage{
add_ends(
  pj,
  segs = "a",
  ends = "e",
  mark = ">",
  size = 1,
  fc = "white",
  fa = 0.5,
  lc = "black",
  la = 0.5,
  lw = 1,
  shape = ".",
  name = "."
)
}
\arguments{
\item{pj}{an object of class \code{'pj'}.}

\item{mark}{character scalar giving the type of end mark to plot. \code{'>'}
indicates a line-only arrowhead end mark. \code{'t'} indicates an
outlineable and fillable (triangular) arrrowhead end mark.\code{'c'}
indicates an outlineable and fillable circular end mark, \code{'d'}
indicates an outlineable and fillable diamond end mark. Finally \code{'s'}
indicates an outlineable and fillable square end mark, and}

\item{size}{positive numeric vector indicating symbol size relative to the
default (thus 1 indicates default sized, 1/2 indicates half sized, and 2
indicates double sized). For characters as marks, 1 indicates point size
10.}

\item{fc}{\code{NA} or a character vector containing valid color
representations giving the color used to fill end marks.}

\item{fa}{\code{NA} or a numeric proportion vector (i.e., between 0 and 1
inclusive) indicating the alpha opacity of the arrow-mark fill color (0 =
fully transparent, 0.5 = translucent, 1 = fully opaque).}

\item{lc}{\code{NA} or a character vector containing valid color
representations giving the color of lines used to draw end marks.}

\item{la}{\code{NA} or a proportion vector (i.e., between 0 and 1 inclusive)
indicating the alpha opacity of the lines used to draw arrow marks.}

\item{lw}{\code{NA} or a positive numeric vector indicating proportional line
width relative to the default giving the width of lines used to draw arrow
marks.}

\item{shape}{character scalar indicating the type of shape to which the
look will be applied.}

\item{loc}{character scalar giving the location of end marks on line
segments. \code{'s'} indicates an end mark at the starts of segments,
\code{'e'} indicates an end mark at the ends of segments, \code{'b'}
indicates end marks at both ends of segments, and \code{'m'} indicates end
marks at the midpoints of segments.}

\item{which}{character scalar indicating which segments in a series of x-y
pairs should be given end marks. \code{'f'} indicates the first segment
only, \code{'l'} indicates the last segment only, \code{'b'} indicates
first and last segments only and \code{'a'} indicates all segments.}
}
\value{
\code{pj} with the addition of the end-mark look applied to the
specified shape.
}
\description{
An ends look is an end mark at the start, midpoint, and/or end
of a line segment. End marks can be line-based arrows, triangle-based
arrows, circles, diamonds, and squares.
}
\section{Valid combinations of argument properties}{
 Intervals can be validly
described using a number of patterns of argument properties as follows,
where \code{base} indicates that the argument set \code{(loc, which, size)}
have all been supplied, \code{line} indicates that the argument set
\code{(lc, la, lw)} have all been supplied, and \code{fill} indicates that
the argument set \code{(fc, fa)} have both been supplied. \describe{
\item{\code{base + mark = ">" + line}}{line-based arrowheads as end
mark.}
\item{\code{base + mark = "t" + line}}{empty triangle-based arrowheads as
end mark.}
\item{\code{base + mark = "t" + fill}}{solid triangle-based arrowheads as
end mark.}
\item{\code{base + mark = "t" + line + fill}}{filled and outlined
triangle-based arrowheads as end mark.}
\item{\code{base + mark = "c" + line}}{empty circle as end mark.}
\item{\code{base + mark = "c" + fill}}{solid circle as end mark.}
\item{\code{base + mark = "c" + line + fill}}{filled and outlined circle
as end mark.}
\item{\code{base + mark = "d" + line}}{empty diamond as end mark.}
\item{\code{base + mark = "d" + fill}}{solid diamond as end mark.}
\item{\code{base + mark = "d" + line + fill}}{filled and outlined diamond
as end mark.}
\item{\code{base + mark = "s" + line}}{empty square as end mark.}
\item{\code{base + mark = "s" + fill}}{solid square as end mark.}
\item{\code{base + mark = "s" + line + fill}}{filled and outlined square
as end mark.}
}
}

\section{Argument recycling}{
 The arguments \code{mark}, \code{size},
\code{lc}, \code{la}, \code{lw}, \code{fc}, and/or \code{fa} are recycled
with the number of segments between x-y pairs of the shape to which the
end-mark look is applied.
}

\seealso{
Other add_look: 
\code{\link{add_ci}()},
\code{\link{add_fills}()},
\code{\link{add_glyphs}()},
\code{\link{add_labels}()},
\code{\link{add_lines}()},
\code{\link{add_marks}()}
}
\concept{add_look}
