% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_look.R
\name{add_fills}
\alias{add_fills}
\title{Add fill looks to a shape.}
\usage{
add_fills(pj, fc = "white", fa = 0.5, shape = ".", name = ".")
}
\arguments{
\item{pj}{an object of class \code{'pj'}.}

\item{fc}{character vector containing valid color representations. Length
must be consistent with the shape a fill look is being added to.}

\item{fa}{proportion vector (i.e., between 0 and 1 inclusive) indicating the
alpha opacity of the fill color (0 = fully transparent, 0.5 = translucent,
1 = fully opaque).}

\item{shape}{character scalar indicating the type of shape to which the
look will be applied.}
}
\value{
\code{pj} with the addition of the fill look applied to the specified
shape.
}
\description{
Not applicable for shapes of type \code{'refline'} or
\code{'segment'}. Adding a fill look to a point shape will effectively turn
the point shape into a free polygon.
}
\section{Argument recycling}{
 The arguments \code{fc} and \code{fa} are
recycled with the number of polygons contained by the shape to which the
fill look is applied.
}

\seealso{
Other add_look: 
\code{\link{add_ci}()},
\code{\link{add_ends}()},
\code{\link{add_glyphs}()},
\code{\link{add_labels}()},
\code{\link{add_lines}()},
\code{\link{add_marks}()}
}
\concept{add_look}
