% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_look.R
\name{add_glyphs}
\alias{add_glyphs}
\title{Add a glyphs look to a shape.}
\usage{
add_glyphs(
  pj,
  glyph = ".",
  size = 10,
  gc = "black",
  ga = 0.5,
  et = NA,
  ec = NA,
  ea = NA,
  shape = ".",
  name = "."
)
}
\arguments{
\item{pj}{an object of class \code{'pj'}.}

\item{glyph}{character vector of glyphs to mark x-y coordinates. Each glyph
must be a single character (which is interpreted literally). The only
exception is when \code{mark = "."}, in which case, a small dot is plotted.}

\item{size}{positive numeric vector giving glyph text size in points.}

\item{gc}{\code{NA} or a character vector indicating the glyph text color.}

\item{ga}{\code{NA} or a proportion vector (i.e., between 0 and 1 inclusive)
indicating the alpha opacity of glyph text colors. 0 = fully transparent,
0.5 = translucent, and 1 = fully opaque.}

\item{et}{character vector indicating the type of enclosure with which to
surround symbols in \code{mark}. Values of \code{NA} indicates no
enclosure. Values of \code{'c'}, \code{'d'} or \code{'s'} indicate an empty
circle, diamond, or square, respectively. \code{'C'}, \code{'D'} or
\code{'S'} indicate a solid circle, diamond, or square, respectively.}

\item{ec}{\code{NA} or a character vector indicating the colors of (optional)
mark enclosures.}

\item{ea}{\code{NA} or a  proportion vector (i.e., between 0 and 1 inclusive)
indicating the alpha opacity of (optional) mark enclosure colors. 0 = fully
transparent, 0.5 = translucent, and 1 = fully opaque.}

\item{shape}{character scalar indicating the type of shape to which the
look will be applied.}
}
\value{
\code{pj} with the addition of the glyph look applied to the
specified shape.
}
\description{
A glyph is a single character. A glyphs look is a glyph marker
for each specific x-y location of a shape. Not applicable for shapes of
type \code{'refline'}.
}
\section{Valid combinations of argument properties}{
 Glyphs can be validly
described using two patterns of argument properties as follows,
where \code{base} indicates that the argument set \code{{mark, size, lc,
  la}} have all been supplied and \code{enclosure} indicates that the
argument set \code{{et, ec, ea}} have all been supplied. \describe{
\item{\code{base}}{The glyph in \code{mark} is printed without any
enclosure}.
\item{\code{base + enclosure}}{The glyph in \code{mark} is printed within
an enclosure}.}
}

\section{Argument recycling}{
 The arguments \code{mark}, \code{size},
\code{lc}, \code{la}, \code{lw}, \code{et}, \code{ec}, and/or \code{ea} are
recycled with the number of x-y pairs used to draw the shape to which the
glyph look is applied
}

\seealso{
Other add_look: 
\code{\link{add_ci}()},
\code{\link{add_ends}()},
\code{\link{add_fills}()},
\code{\link{add_labels}()},
\code{\link{add_lines}()},
\code{\link{add_marks}()}
}
\concept{add_look}
