% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_look.R
\name{add_labels}
\alias{add_labels}
\title{Add a labels look to a shape.}
\usage{
add_labels(
  pj,
  text,
  size = 10,
  font = "sans",
  style = "plain",
  tc = "black",
  ta = 0.5,
  hj = 0.5,
  vj = 0.5,
  parse = FALSE,
  a = 0,
  fc = "white",
  fa = 0.5,
  lc = "black",
  la = 0.5,
  lt = "-",
  lw = 1,
  au = "r",
  dx = 0,
  dy = 0,
  shape = ".",
  name = "."
)
}
\arguments{
\item{pj}{an object of class \code{'pj'}.}

\item{text}{character vector containing labels to add to x-y locations of
shapes. Special values are \code{'@xX'} for printing the x values with
\code{X} decimal point; \code{'@yY'} for printing the y values with
\code{Y} decimal points, and \code{'@xXyY'} to print the x and y values
with \code{X} and \code{Y} decimal points respectively. \cr\cr \code{X} and
\code{Y} can be any digit from 0 to 9. For example, if the x and y values
of a point were \code{pi} and \code{e}, the following labels wquld be
printed at the coordinate \code{(pi, e)}: \code{'@x0'} would print
\code{'3'}, \code{'@y8'} would print \code{'2.71828183'}, and
\code{'@x2y3'} would print \code{'(3.14, 2.718)'}}

\item{size}{positive numeric vector giving text size in points.}

\item{font}{character vector giving label fonts. Valid values are
\code{'mono'}, \code{'sans'}, \code{'serif'}, \code{'American Typewriter'},
\code{'Arial'}, \code{'Bradley Hand ITC'}, \code{'Calibri'}, \code{'Comic
  Sans MS'}, \code{'Courier'}, \code{'Courier New'}, \code{'Garamond'},
\code{'Geneva'}, \code{'Helvetica'}, \code{'Impact'}, \code{'Monaco'},
\code{'Palatino'}, \code{'Symbol'}, \code{'Tahoma'}, \code{'Times'},
\code{'Times New Roman'}, \code{'Trebuchet MS'}, and \code{'Verdana'}.
\cr\cr \code{'mono'} uses the default monospaced font for the device/R
configuration. \code{'sans'} uses the default sans serif font for the
device/R configuration. \code{'serif'} uses the default serif font for the
device/R configuration.}

\item{style}{character vector giving text style. \code{'p'} indicates plain,
\code{'b'} indicates bold, \code{'i'} indicates italic, and \code{'bi'}
indicates bold italic.}

\item{tc}{character vector giving valid text color specifications for labels.}

\item{ta}{proportion vector (i.e., between 0 and 1 inclusive) indicating the
alpha opacity of the text color (0 = fully transparent, 0.5 = translucent,
1 = fully opaque).}

\item{hj, vj}{numeric proportion vectors (i.e., all values are between 0 and 1
inclusive) giving the horizontal justification and vertical justification,
respectively. \code{0} indicates left or bottom justification, \code{1}
indicates right or top justification, and \code{0.5} indicates middle or
center justification.}

\item{parse}{logical vector indicating whether to parse the labels (e.g., to
print labels with Greek letters or mathematical expressions).}

\item{a}{numeric vector giving angle of text rotation.}

\item{fc}{\code{NA} or a character vector of fill colors for (optional) label
background rectangles.}

\item{fa}{\code{NA} or a proportion vector (i.e., all values are between 0
and 1, inclusive) indicating fill alpha opacity values for (optional) label
background rectangles (0 is transparent, 1 is opaque, 0.5 is translucent).}

\item{lc}{\code{NA} or a character vector of line colors for (optional) label
background rectangles.}

\item{la}{\code{NA} or a proportion vector (i.e., all values are between 0
and 1, inclusive) indicating line alpha opacity values for (optional) label
background rectangles (0 is transparent, 1 is opaque, 0.5 is translucent).}

\item{lt}{\code{NA} or a character vector indicating line type for (optional)
label background rectangles, where \code{'-'} indicates a solid line,
\code{':'} indicates a dotted line, \code{'='} indicates a dashed line,
\code{':='} indicates a dot-dash line, \code{'LD'} indicates a long-width
dashed line, and \code{'2D'} indicates a double-width dashed line.}

\item{lw}{\code{NA} or a positive numeric vector indicating proportional line
width for (optional) background rectangles relative to the default (i.e.,
\code{1} gives the default line width, 0\code{1/2} gives half the default
line width, and \code{2} gives double the default line width.)}

\item{au}{\code{NA} or a character vector giving units of angle of rotation.
\code{'d'} indicates degrees, \code{'g'} indicates gradians, \code{'r'}
indicates radians, and \code{'p'} indicates proportions of a full
revolution.}

\item{dx}{numeric vector giving horizontal offset from \code{x} for label
positioning.}

\item{dy}{numeric vector giving vertical offset from \code{x} for label
positioning.}

\item{shape}{character scalar indicating the type of shape to which the
look will be applied.}
}
\value{
\code{pj} with the addition of the label look applied to the
specified shape.
}
\description{
Not applicable for shapes of type \code{'refline'}.
}
\section{Valid combinations of argument properties}{
 Labels can be validly
described using patterns of argument properties as follows, where
\code{base} indicates that the argument set \code{(text, tc, ta, font,
  size, style, hj, vj, a, parse, au, dx, dy)} have all been supplied,
\code{line} indicates that the argument set \code{(lc, la, lt)} have all
been supplied. and \code{fill} indicates that the argument set \code{(fc,
  fa)} have both been supplied. \describe{
\item{\code{base}}{The label \code{text} is printed without any
enclosure}.
\item{\code{base + line}}{The label in \code{text} is printed in an
outlined enclosure}.
\item{\code{base + fill}}{The label in \code{text} is printed in a filled
enclosure}.
\item{\code{base + line + fill}}{The label in \code{text} is printed in
an outlined and filled enclosure}.}
}

\section{Argument recycling}{
 The arguments \code{text}, \code{tc}, \code{ta},
\code{font}, \code{size}, \code{style}, \code{hj}, \code{vj}, \code{a},
\code{parse}, \code{fc}, \code{fa}, \code{lc}, \code{la}, \code{lt},
\code{lw}, \code{au}, \code{dx}, and/or \code{dy} are recycled with the
number of \code{x-y} pairs used to draw the shape to which the label look
is applied
}

\seealso{
Other add_look: 
\code{\link{add_ci}()},
\code{\link{add_ends}()},
\code{\link{add_fills}()},
\code{\link{add_glyphs}()},
\code{\link{add_lines}()},
\code{\link{add_marks}()}
}
\concept{add_look}
