% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_look.R
\name{add_lines}
\alias{add_lines}
\title{Add a lines look to a shape.}
\usage{
add_lines(
  pj,
  lc = "black",
  la = 0.5,
  lt = "-",
  lw = 1,
  shape = ".",
  name = "."
)
}
\arguments{
\item{pj}{an object of class \code{'pj'}.}

\item{lc}{character vector containing valid color representations. Length
must be consistent with the shape a line look is being added to.}

\item{la}{proportion vector (i.e., between 0 and 1 inclusive) indicating the
alpha opacity of the line color (0 = fully transparent, 0.5 = translucent,
1 = fully opaque).}

\item{lt}{character vector indicating line type, where \code{'-'} indicates a
solid line, \code{':'} indicates a dotted line, \code{'='} indicates a
dashed line, \code{':='} indicates a dot-dash line, \code{'LD'} indicates a
long-width dashed line, and \code{'2D'} indicates a double-width dashed
line.}

\item{lw}{positive numeric vector indicating proportional line width relative
to the default (i.e., \code{1} gives the default line width, \code{1/2}
gives half the default line width, and \code{2} gives double the default
line width.)}

\item{shape}{character scalar indicating the type of shape to which the
look will be applied.}
}
\value{
\code{pj} with the addition of the line look applied to the
specified shape.
}
\description{
Add a lines look to a shape.
}
\section{Argument recycling}{
 For shapes that are polygons, the arguments
\code{lc}, \code{la}, \code{lt}, and \code{lw} are recycled with the number
of distinct polygons described by the shape to which the look is to be
applied. For point shapes, these arguments are recycled with the number of
defined points. For refline and segment shapes, the arguments must be
recyclable with the number of lines/segments defined by the shape.
}

\seealso{
Other add_look: 
\code{\link{add_ci}()},
\code{\link{add_ends}()},
\code{\link{add_fills}()},
\code{\link{add_glyphs}()},
\code{\link{add_labels}()},
\code{\link{add_marks}()}
}
\concept{add_look}
