% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_look.R
\name{add_marks}
\alias{add_marks}
\title{Add a marks look to a shape.}
\usage{
add_marks(
  pj,
  mark = "d",
  size = 1,
  fc = "white",
  fa = 0.5,
  lc = "black",
  la = 0.5,
  lw = 1,
  et = NA,
  ec = NA,
  ea = NA,
  shape = ".",
  name = "."
)
}
\arguments{
\item{pj}{an object of class \code{'pj'}.}

\item{mark}{character vector of symbols to mark x-y coordinates. Each value
must be a single character, where \code{'c'} indicates a circle, \code{'d'}
indicates a diamond, \code{'s'} indicates a square, \code{'3'} through
\code{'9'} indicate 3 through 9 pointed stars, \code{'<'} indicates a
left-pointing triangle,  \code{'>'} indicates a right-pointing triangle,
\code{'v'} indicates a down-pointing triangle, and \code{'^'} indicates an
up-pointing triangle.}

\item{size}{positive numeric vector indicating symbol size relative to the
default (thus 1 indicates default sized, 1/2 indicates half sized, and 2
indicates double sized).}

\item{et}{character vector indicating the type of enclosure with which to
surround symbols in \code{mark}. Values of \code{NA} indicates no
enclosure. Values of \code{'c'}, \code{'d'} or \code{'s'} indicate an empty
circle, diamond, or square, respectively. \code{'C'}, \code{'D'} or
\code{'S'} indicate a solid circle, diamond, or square, respectively.}

\item{ec}{\code{NA} or a character vector indicating the colors of (optional)
mark enclosures.}

\item{ea}{\code{NA} or a  proportion vector (i.e., between 0 and 1 inclusive)
indicating the alpha opacity of (optional) mark enclosure colors. 0 = fully
transparent, 0.5 = translucent, and 1 = fully opaque.}

\item{shape}{character scalar indicating the type of shape to which the
look will be applied.}
}
\value{
\code{pj} with the addition of the mark look applied to the
specified shape.
}
\description{
A mark is a polygon-based marker for a given x-y location. Not
applicable for shapes of type \code{'refline'}.
}
\section{Valid combinations of argument properties}{
 Marks can be validly
described using the patterns of argument properties as follows, where
\code{base} indicates that the argument set \code{{mark, size}} has been
supplied, \code{fill} indicates that the argument set \code{{fc, fa}} has
been supplied, \code{line} inidicates that the argument set \code{{lc, la,
  lt, lw}} have been supplied and \code{enclosure} indicates that the
argument set \code{{et, ec, ea}} has been supplied. \describe{
\item{\code{base + line}}{Outlined mark.}
\item{\code{base + fill}}{Solid mark.}
\item{\code{base + line + fill}}{Outlined and filled mark}.
\item{\code{base + line + enclosure}}{Outlined mark in an enclosure}.
\item{\code{base + fill + enclosure}}{Solid mark in an enclosure}.
\item{\code{base + line + fill + enclosure}}{Outlined and filled mark in
an enclosure}.
}
}

\section{Argument recycling}{
 The arguments \code{mark}, \code{size},
\code{lc}, \code{la}, \code{lw}, \code{fc}, \code{fa}, \code{et},
\code{ec}, and/or \code{ea} are recycled with the number of x-y pairs used
to draw the shape to which the mark look is applied.
}

\seealso{
Other add_look: 
\code{\link{add_ci}()},
\code{\link{add_ends}()},
\code{\link{add_fills}()},
\code{\link{add_glyphs}()},
\code{\link{add_labels}()},
\code{\link{add_lines}()}
}
\concept{add_look}
