% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_calc_shape_xy.R
\name{int_calc_shape_xy}
\alias{int_calc_shape_xy}
\alias{int_calc_point_xy}
\alias{int_calc_segment_xy}
\alias{int_calc_refline_xy}
\alias{int_calc_area_xy}
\alias{int_calc_ribbon_xy}
\alias{int_calc_triangle_xy}
\alias{int_calc_isosceles_xy}
\alias{int_calc_right_xy}
\alias{int_calc_quad_xy}
\alias{int_calc_rect_xy}
\alias{int_calc_diamond_xy}
\alias{int_calc_kite_xy}
\alias{int_calc_polygon_xy}
\alias{int_calc_regpoly_xy}
\alias{int_calc_star_xy}
\alias{int_calc_freepoly_xy}
\alias{int_calc_circle_xy}
\alias{int_calc_ellipse_xy}
\alias{int_calc_round_xy}
\title{Calculate x-y coordinates for shapes}
\usage{
int_calc_point_xy(x)

int_calc_segment_xy(x)

int_calc_refline_xy(x, rg)

int_calc_area_xy(x, rg)

int_calc_ribbon_xy(x)

int_calc_triangle_xy(x)

int_calc_isosceles_xy(x)

int_calc_right_xy(x)

int_calc_quad_xy(x)

int_calc_rect_xy(x)

int_calc_diamond_xy(x)

int_calc_kite_xy(x)

int_calc_polygon_xy(x)

int_calc_regpoly_xy(x)

int_calc_star_xy(x)

int_calc_freepoly_xy(x)

int_calc_circle_xy(x)

int_calc_ellipse_xy(x)

int_calc_round_xy(x)
}
\arguments{
\item{x}{A tibble row defining an instance of a shape.}

\item{rg}{A tibble row defining the region a shape is attached to.}
}
\description{
Calculate x-y coordinates for shapes
}
