% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_new_id.R
\name{int_new_id}
\alias{int_new_id}
\title{Create new region, shape, look, or mod ID numbers}
\usage{
int_new_id(jp, type)
}
\arguments{
\item{jp}{A jplot object.}

\item{type}{Character scalar from \code{c('region', 'shape', 'look', 'mod')}.}
}
\value{
Non-negative whole number scalar.
}
\description{
Create new region, shape, look, or mod ID numbers
}
