% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mod.R
\name{mod_dilate}
\alias{mod_dilate}
\title{Add a dilation modification to a shape}
\usage{
mod_dilate(pj, x, y, mx, my, shape = ".", name = ".")
}
\arguments{
\item{pj}{an object of class \code{'pj'}.}

\item{x}{integer vector identifying vertical fixed points for anchoring
shapes in dilation.}

\item{mx}{positive numeric vector giving the value by which to multiply shape
width (x direction).}

\item{my}{positive numeric vector giving the value by which to multiply shape
height (y direction).}

\item{shape}{character scalar indicating the type of shape to which the
mod will be applied.}

\item{name}{character scalar name to give to the specified mod. The special
value \code{"."} indicates that the pj package should create a default name
for the mod.}
}
\value{
\code{pj} with the dilation modification applied to the specified
shape.
}
\description{
A dilation modification squeezes and/or stretches a shape in the
x and y directions. Dilations are not valid for refline shapes.
}
\section{Argument recycling}{
 The argument set \code{{x, y, mx, my, xu, yu}}
is recycled with the number of subshapes defined for a given shape. For
point, area, and ribbon shapes, all arguments in the set must be scalar.
For segment shapes, arguments in the set must be recyclable with the number
of segments defined by that shape. For shapes that define polygons,
arguments in the set must be recyclable with the number of polygons defined
by the shape.
}

\seealso{
Other add_mod: 
\code{\link{mod_reflect}()},
\code{\link{mod_rotate}()},
\code{\link{mod_translate}()}
}
\concept{add_mod}
