% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mod.R
\name{mod_reflect}
\alias{mod_reflect}
\title{Add a reflection modification to a shape}
\usage{
mod_reflect(
  pj,
  ref = "h",
  x1 = NA,
  y1 = NA,
  x2 = NA,
  y2 = NA,
  s = NA,
  i = NA,
  a = NA,
  au = "r",
  shape = ".",
  name = "."
)
}
\arguments{
\item{pj}{an object of class \code{'pj'}.}

\item{ref}{\code{NA} or character vector identifying pre-defined reference
lines. See the \emph{the ref argument} section.}

\item{x1}{\code{NA} or numeric vectors indicating either a vertical reference
line at the value contained in \code{x1} (if used alone) or the horizontal
location of a primary point used to define a reference line (if used with
\code{y1}).}

\item{y1}{\code{NA} or numeric vectors indicating either a horizontal
reference line at the value contained in\code{y1} (if used alone) or the
vertical location of a primary point used to define a reference line (if
used with \code{x1}).}

\item{x2}{\code{NA} or numeric vector indicating the vertical locations of
the secondary points to define lines in conjunction with \code{x1},
\code{y1}, and \code{y2}.}

\item{s}{\code{NA} or numeric vector indicating the slope of the reference
line.}

\item{i}{\code{NA} or numeric vector indicating the x-intercept of the
reference line.}

\item{a}{\code{NA} or numeric vector indicating the angle of the reference
line.}

\item{au}{\code{NA} or character vector indicating the angle units to be used
where \code{'d'} indicates degrees, \code{'g'} indicates gradians,
\code{'p'} indicates proportion of a full revolution, and \code{'r'}
indicates radians.}

\item{shape}{character scalar indicating the type of shape to which the
mod will be applied.}

\item{name}{character scalar name to give to the specified mod. The special
value \code{"."} indicates that the pj package should create a default name
for the mod.}
}
\value{
\code{pj} with the reflection modification applied to the specified
shape.
}
\description{
A reflection modification reflects a shape around a reference
line. Reflections are not valid for refline shapes.
}
\section{The ref argument}{

\itemize{
\item \code{'i'} indicates reflection across the identity line,
\item \code{'x'} indicates reflection across the x-axis,
\item \code{'y'} indicates reflection across the y-axis,
\item \code{'l'} indicates reflection across the vertical line through
the left edge of a shape's bounding rectangle,
\item \code{'r'} indicates reflection across the vertical line through
the right edge of a shape's bounding rectangle,
\item \code{'b'} indicates reflection across the horizontal line through
the bottom edge of a shape's bounding rectangle,
\item \code{'t'} indicates reflection across the horizontal line through
the top edge of a shape's bounding rectangle,
\item \code{'h'} indicates horizontal reflection across the vertical
midline of the shape's bounding rectangle.
\item \code{'v'} indicates vertical reflection across the horizontal
midline of the shape's bounding rectangle.
\item \code{'d'} indicates reflection across the line through th
downward diagonal between the top-left and bottom-right
corners of a shape's bounding rectangle, and
\item \code{'u'} indicates reflection across the line through the upward
diagonal between to bottom-left and top-right corners of
a shape's bounding rectangle.
}
}

\section{Defining reference lines}{
 Valid combinations of non-\code{NA}
locating arguments are given below. \describe{
\item{\code{ref}}{Pre-defined reference line}
\item{\code{x1}}{Vertical line at \code{x1}.}
\item{\code{y1}}{Horizontal line at \code{y1}.}
\item{\code{x1 + y1 + x2 + y2}}{Point-point definition (line through
\code{(x1, y1)} and \code{(x2, y2)}).}
\item{\code{x1 + y1 + a}}{Point-angle definition (line through
\code{(x1, y1)} at angle \code{a}).}
\item{\code{x1 + y1 + i}}{Point-intercept definition (line through
intercept and \code{(x1, y1)}).}
\item{\code{x1 + y1 + s}}{Point-slope definition (line through
\code{(x1, y1)} with slope \code{s}).}
\item{\code{a + i}}{Angle-intercept definition (line through intercept at
angle \code{a}).}
\item{\code{s + i}}{Slope-intercept definition (line through intercept
with slope \code{s}).}}
}

\section{Argument recycling}{
 The argument set \code{{x, y, ref, s, i, x2, y2,
  a, au, xu, yu}} is recycled with the number of subshapes defined for a
given shape. For point, area, and ribbon shapes, all arguments in the set
must be scalar. For segment shapes, arguments in the set must be recyclable
with the number of segments defined by that shape. For shapes that define
polygons, arguments in the set must be recyclable with the number of
polygons defined by the shape.
}

\seealso{
Other add_mod: 
\code{\link{mod_dilate}()},
\code{\link{mod_rotate}()},
\code{\link{mod_translate}()}
}
\concept{add_mod}
