% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mod.R
\name{mod_rotate}
\alias{mod_rotate}
\title{Add a rotation modification to a shape}
\usage{
mod_rotate(
  pj,
  ref = "c",
  x = NA,
  y = NA,
  a = NA,
  au = "r",
  shape = ".",
  name = "."
)
}
\arguments{
\item{pj}{an object of class \code{'pj'}.}

\item{ref}{\code{NA} or a character vector identifying pre-defined
reference points. See the \emph{the ref argument} section.}

\item{x}{\code{NA} or a numeric vector indicating the horizontal location of
a point around which to rotate shapes.}

\item{y}{\code{NA} or a numeric vector indicating the vertical location of a
point around which to rotate shapes.}

\item{a}{\code{NA} or a numeric vector indicating the angle of
rotation.}

\item{au}{\code{NA} or a character vector indicating the angle
units to be used where \code{'d'} indicates degrees, \code{'g'} indicates
gradians, \code{'p'} indicates proportion of a full revolution, and
\code{'r'} indicates radians.}

\item{shape}{character scalar indicating the type of shape to which the
mod will be applied.}

\item{name}{character scalar name to give to the specified mod. The special
value \code{"."} indicates that the pj package should create a default name
for the mod.}
}
\value{
\code{pj} with the rotation modification applied to the specified
shape.
}
\description{
A rotation modification rotates a shape around a reference
point. Rotations are not valid for refline shapes.
}
\section{The ref argument}{

\itemize{
\item \code{'c'} indicates center point of a shape's bounding rectangle.
\item \code{'o'} indicates the origin (0, 0).
\item \code{'bl'} indicates the bottom-left corner of a shape's bounding
rectangle.
\item \code{'br'} indicates the bottom-right corner of a shape's bounding
rectangle.
\item \code{'tl'} indicates the top-left corner of a shape's bounding
rectangle.
\item \code{'tr'} indicates the top-right corner of a shape's bounding
rectangle.
\item \code{'lm'} indicates the mid-point of the left edge of a shape's
bounding rectangle.
\item \code{'rm'} indicates the mid-point of the right edge of a shape's
bounding rectangle.
\item \code{'bm'} indicates the mid-point of the bottom edge of a shape's
bounding rectangle.
\item \code{'tm'} indicates the mid-point of the top edge of a shape's
bounding rectangle.
}
}

\section{Argument recycling}{
 The argument set \code{{ref, x, y, a, au, xu,
  yu}} is recycled with the number of subshapes defined for a given shape.
For point, area, and ribbon shapes, all arguments in the set must be
scalar. For segment shapes, arguments in the set must be recyclable with
the number of segments defined by that shape. For shapes that define
polygons, arguments in the set must be recyclable with the number of
polygons defined by the shape.
}

\seealso{
Other add_mod: 
\code{\link{mod_dilate}()},
\code{\link{mod_reflect}()},
\code{\link{mod_translate}()}
}
\concept{add_mod}
