% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mod.R
\name{mod_translate}
\alias{mod_translate}
\title{Add a translation modification to a shape}
\usage{
mod_translate(pj, dx, dy, shape = ".", name = ".")
}
\arguments{
\item{pj}{an object of class \code{'pj'}.}

\item{dx}{numeric vector indicating the distances for horizontal translation.}

\item{dy}{numeric vector indicating the distances for vertical translation.}

\item{shape}{character scalar indicating the type of shape to which the
mod will be applied.}

\item{name}{character scalar name to give to the specified mod. The special
value \code{"."} indicates that the pj package should create a default name
for the mod.}
}
\value{
\code{pj} with the translation modification applied to the specified
shape.
}
\description{
A translation modification moves a shape in the x and/or y
directions.
}
\section{Argument recycling}{
 The argument set \code{{x, y, xu, yu}} is
recycled with the number of subshapes defined for a given shape. For point,
area, and ribbon shapes, all arguments in the set must be scalar. For
segment shapes, arguments in the set must be recyclable with the number of
segments defined by that shape. For shapes that define polygons, arguments
in the set must be recyclable with the number of polygons defined by the
shape.
}

\seealso{
Other add_mod: 
\code{\link{mod_dilate}()},
\code{\link{mod_reflect}()},
\code{\link{mod_rotate}()}
}
\concept{add_mod}
