% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_polycurve.R
\name{new_area}
\alias{new_area}
\title{Add an area shape to a plot}
\usage{
new_area(
  pj,
  x,
  y,
  ref = "x",
  dir = "v",
  x1 = NA,
  y1 = NA,
  x2 = NA,
  y2 = NA,
  s = NA,
  i = NA,
  a = NA,
  au = "r",
  region = ".",
  look = NULL,
  mod = NULL,
  name = "."
)
}
\arguments{
\item{pj}{an object of class \code{'pj'}.}

\item{x}{numeric vector defining the horizontal locations of points on a
curve. The longer of \code{x} and \code{y} must have at least 2 elements.}

\item{y}{numeric vector defining the vertical locations of points on a curve.
The longer of \code{x} and \code{y} must have at least 2 elements.}

\item{ref}{\code{NA} or a character scalar giving a pre-defined reference
line. \code{'i'} indicates the identity line \code{'x'} indicates the
x-axis, and \code{'y'} indicates the y-axis.}

\item{dir}{character scalar identifying the directional orientation of the
reference line, where \code{'v'} indicates a vertically-oriented area plot
(the vertical area between the x-y curve and the reference line) and
\code{'h'} indicates a horizontally-oriented area plot.}

\item{x1}{\code{NA} or a numeric scalar indicating either a vertical
reference line at the value contained in \code{x1} (if used alone) or the
horizontal location of a primary point used to define a reference line (if
used with \code{y1}).}

\item{y1}{\code{NA} or a numeric scalar indicating either a horizontal
reference line at the value contained in\code{y1} (if used alone) or the
vertical location of a primary point used to define a reference line (if
used with \code{x1}).}

\item{x2}{\code{NA} or a numeric scalar indicating the horizontal location of
the secondary points to define a reference line (in conjunction with
\code{x1}, \code{y1}, and \code{y2}).}

\item{y2}{\code{NA} or a numeric scalar indicating the vertical location of
the secondary points to define a reference line (in conjunction with
\code{x1}, \code{y1}, and \code{x2}).}

\item{s}{\code{NA} or a numeric scalar indicating the slope of the reference
line.}

\item{i}{\code{NA} or a numeric scalar indicating the intercept of the
reference line.}

\item{a}{\code{NA} or a numeric scalar indicating the angle of the reference
line.}

\item{au}{\code{NA} or a character vector indicating the angle units to be
used where \code{'d'} indicates degrees, \code{'g'} indicates gradians,
\code{'p'} indicates proportion of a full revolution, and \code{'r'}
indicates radians.}

\item{region}{either character scalar \code{"."} to indicate the most
recently used region, a positive integer to indicate the ID number of a
region (\code{0} indicates the background region), or a character scalar to
indicate the name of a region (\code{'bg'} indicates the background
region).}

\item{look}{an optional list containing up to four lists specifying looks to
apply to the shape. See the \emph{using \code{look} to make shapes appear}
section.}

\item{mod}{an optional list containing up to four lists specifying
modifications to apply to the shape. See the \emph{using \code{mod} to
transform shapes} section.}

\item{name}{character scalar indicating a name for the shape. The special
value \code{"."} indicates that the pj package create a default name for
the shape.}
}
\value{
\code{pj} with the addition of the specified area shape.
}
\description{
An area shape is a polygon defined as the area between a curve
(a set of x-y pairs) and a reference line. The default reference line is
the x-axis.
}
\section{Defining reference lines}{
 Valid combinations of non-\code{NA}
locating arguments are given below. \describe{
\item{\code{ref}}{Pre-defined reference line}
\item{\code{x1}}{Vertical line at \code{x1}.}
\item{\code{y1}}{Horizontal line at \code{y1}.}
\item{\code{x1 + y1 + x2 + y2}}{Point-point definition (line through
\code{(x1, y1)} and \code{(x2, y2)}).}
\item{\code{x1 + y1 + a}}{Point-angle definition (line through
\code{(x1, y1)} at angle \code{a}).}
\item{\code{x1 + y1 + i}}{Point-intercept definition (line through
intercept and \code{(x1, y1)}).}
\item{\code{x1 + y1 + s}}{Point-slope definition (line through
\code{(x1, y1)} with slope \code{s}).}
\item{\code{a + i}}{Angle-intercept definition (line through intercept
at angle \code{a}).}
\item{\code{s + i}}{Slope-intercept definition (line through intercept
with slope \code{s}.}}
}

\section{Argument recycling}{
 The argument set \code{\{x, y\}} is recycled.
}

\section{Using \code{look} to make shapes appear}{
 By itself, this function
simply defines the location of x-y pairs in the assigned region for the
specified shape. Various looks can be added to the shape using the
\code{look} parameter or via the \code{\link{add_ends}},
\code{\link{add_fills}}, \code{\link{add_glyphs}},
\code{\link{add_labels}}, \code{\link{add_lines}}, and
\code{\link{add_marks}} functions. Sub-lists of the look argument must be
named \code{'end'}, \code{'fill'}, \code{'glyph'}, \code{'label'},
\code{'line'}, and/or \code{'mark'} to indicate that line end marks, fill
colors, glyphs (single characters), labels, lines, and/or point marks be
added to the shape.
}

\section{Using \code{mod} to transform shapes}{
 Shapes can be transformed in
four ways: dilation, reflection, rotation, and translation/. using the
\code{mod} argument or the \code{\link{mod_dilate}},
\code{\link{mod_reflect}}, \code{\link{mod_rotate}}, and
\code{\link{mod_translate}}. Sub-lists of the \code{mod} argument must be
named \code{'dilate'}, \code{'reflect'}, \code{'rotate'}, and/or
\code{'translate'} to indicate that a dilation, reflection, rotation, and
or translation be applied to the shape.
}

\seealso{
Other new_polycurve: 
\code{\link{new_ribbon}()}
}
\concept{new_polycurve}
