% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_polys.R
\name{new_free}
\alias{new_free}
\title{Add a free polygon shape to a plot}
\usage{
new_free(pj, x, y, region = ".", look = NULL, mod = NULL, name = ".")
}
\arguments{
\item{pj}{an object of class \code{'pj'}.}

\item{x}{numeric vector defining the horizontal location of each vertex. The
longer of \code{x} and \code{y} must be of length 3 or greater.}

\item{y}{numeric vector defining the vertical location of each vertex. The
longer of \code{x} and \code{y} must be of length 3 or greater.}

\item{region}{either character scalar \code{"."} to indicate the most
recently used region, a positive integer to indicate the ID number of a
region (\code{0} indicates the background region), or a character scalar to
indicate the name of a region (\code{'bg'} indicates the background
region).}

\item{look}{an optional list containing up to four lists specifying looks to
apply to the shape. See the \emph{using \code{look} to make shapes appear}
section.}

\item{mod}{an optional list containing up to four lists specifying
modifications to apply to the shape. See the \emph{using \code{mod} to
transform shapes} section.}

\item{name}{character scalar indicating a name for the shape. The special
value \code{"."} indicates that the pj package create a default name for
the shape.}
}
\value{
\code{pj} with the addition of the specified free polygon shape.
}
\description{
A free polygon shape is a polygon of arbitrary shape defined by
any sequence of x-y pairs where the last x-y pair is connected to the first
x-y pair to close the polygon.
}
\section{Argument recycling}{
 The argument set \code{\{x, y\}} is recycled.
}

\section{Using \code{look} to make shapes appear}{
 By itself, this function
simply defines the location of x-y pairs in the assigned region for the
specified shape. Various looks can be added to the shape using the
\code{look} parameter or via the \code{\link{add_ends}},
\code{\link{add_fills}}, \code{\link{add_glyphs}},
\code{\link{add_labels}}, \code{\link{add_lines}}, and
\code{\link{add_marks}} functions. Sub-lists of the look argument must be
named \code{'end'}, \code{'fill'}, \code{'glyph'}, \code{'label'},
\code{'line'}, and/or \code{'mark'} to indicate that line end marks, fill
colors, glyphs (single characters), labels, lines, and/or point marks be
added to the shape.
}

\section{Using \code{mod} to transform shapes}{
 Shapes can be transformed in
four ways: dilation, reflection, rotation, and translation/. using the
\code{mod} argument or the \code{\link{mod_dilate}},
\code{\link{mod_reflect}}, \code{\link{mod_rotate}}, and
\code{\link{mod_translate}}. Sub-lists of the \code{mod} argument must be
named \code{'dilate'}, \code{'reflect'}, \code{'rotate'}, and/or
\code{'translate'} to indicate that a dilation, reflection, rotation, and
or translation be applied to the shape.
}

\seealso{
Other new_polygons: 
\code{\link{new_polys}()},
\code{\link{new_regulars}()},
\code{\link{new_stars}()}
}
\concept{new_polygons}
