% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_triangle.R, R/new_triangles.R
\name{new_isosceles}
\alias{new_isosceles}
\title{Add one or more isosceles triangle shapes}
\usage{
new_isosceles(
  pj,
  x,
  y,
  b = NA,
  h = NA,
  e = NA,
  re = NA,
  te = NA,
  eb = NA,
  ee = NA,
  loc = "b",
  px = 0.5,
  py = 0.5,
  au = "r",
  region = ".",
  look = NULL,
  mod = NULL,
  name = "."
)

new_isosceles(
  pj,
  x,
  y,
  b = NA,
  h = NA,
  e = NA,
  re = NA,
  te = NA,
  eb = NA,
  ee = NA,
  loc = "b",
  px = 0.5,
  py = 0.5,
  au = "r",
  region = ".",
  look = NULL,
  mod = NULL,
  name = "."
)
}
\arguments{
\item{pj}{an object of class \code{'pj'}.}

\item{x}{numeric vector giving horizontal location. When \code{x} is used in
conjunction with \code{px} and \code{w} these vectors define a fixed
horizontal point for each shape. When \code{x} is used in conjunction with
\code{re}, \code{x} defines the left edge of the bounding rectangle of each
shape.}

\item{y}{numeric vector giving vertical location. When \code{y} is used in
conjunction with \code{py} and \code{h} these vectors define a fixed
vertical point for each shape. When \code{y} is used in conjunction with
\code{te},\code{y} defines the bottom edge of the bounding rectangle of
each shape.}

\item{b}{\code{NA} or a positive numeric vector indicating base edge lengths.}

\item{h}{\code{NA} or a numeric vector defining the height of each bounding
rectangle.}

\item{e}{\code{NA} or a positive numeric vector indicating non-base edge
lengths.}

\item{re}{\code{NA} or a numeric vector defining the right edge of each
bounding rectangle.}

\item{te}{\code{NA} or a numeric vector defining the top edge of each
bounding rectangle.}

\item{ee}{\code{NA} or a positive numeric vector indicating the angles
between the two same-length non-base edges.}

\item{loc}{character vector indicating where the base of the isosceles
triangles are located. \code{'l'}, \code{'r'}, \code{'b'}, and \code{'t'}
indicate left, right, bottom, and top edges of the bounding rectangle,
respectively.}

\item{px}{\code{NA} or a proportion vector defining the location of \code{x}
as the proportion of the distance from left to right edge of bounding
rectangles. Values are generally between 0 and 1, inclusive, however, when
they are outside those values, fixed points \code{x} will lie outside
bounding rectangles.}

\item{py}{\code{NA} or a proportion vector defining the location of \code{y}
as the proportion of the distance from bottom to top edge of bounding
rectangles. Values are generally between 0 and 1, inclusive, however, when
they are outside those values, fixed points \code{y} will lie outside
bounding rectangles.}

\item{au}{\code{NA} or a character vector indicating angle units where
\code{'d'}, \code{'g'}, \code{'r'}, and \code{'p'} indicate degrees,
gradians, radians, and proportions of a revolution.}

\item{region}{either character scalar \code{"."} to indicate the most
recently used region, a positive integer to indicate the ID number of a
region (\code{0} indicates the background region), or a character scalar to
indicate the name of a region (\code{'bg'} indicates the background
region).}

\item{look}{an optional list containing up to four lists specifying looks to
apply to the shape. See the \emph{using \code{look} to make shapes appear}
section.}

\item{mod}{an optional list containing up to four lists specifying
modifications to apply to the shape. See the \emph{using \code{mod} to
transform shapes} section.}

\item{name}{character scalar indicating a name for the shape. The special
value \code{"."} indicates that the pj package create a default name for
the shape.}

\item{s}{(optional) positive numeric vector indicating non-base side lengths.}

\item{sb, ss}{(optional) positive numeric vector indicating, respectively, the
angles between (non-base side) and base side vs. angles between the two
equal-length non-base sides.}

\item{be}{\code{NA} or a positive numeric vector indicating the angles
between (non-base) edges and base edge.}
}
\value{
\code{jp} with the addition of the specified isosceles-triangle
shape(s).

\code{pj} with the addition of the defined isosceles-triangles.
}
\description{
isosceles triangle shapes are located in plotting regions
using bounding rectangles; a concept used in defining function parameters.
See the \emph{locating bounding rectangles} section for details.

An isosceles triangles shapes is one or more isosceles triangles
defined using the same subset of locating parameters (\code{x}, \code{y},
\code{px}, \code{py}, \code{b}, \code{h}, \code{e}, \code{be}, \code{ee},
\code{au}, \code{re}, and \code{te}). Isosceles triangles are located in
plotting regions using bounding rectangles, a concept used in defining
function arguments. See the \emph{defining bounding rectangles} section
for details.
}
\section{The h and s arguments}{
 \code{h}) is calculated as the perpendicular
distance from the vertex between non-base sides to the base side. \code{s}
is defined as the length of the two equivalent-length non-base sides.
}

\section{Locating bounding rectangles}{
 Bounding rectangles can be located in
a number of ways. The following shows valid combination of non-\code{NA}
locating arguments: \describe{
\item{\code{(x + px) + b + (y + py) + h}}{(x-anchor) + base length +
(y-anchor) + height}
\item{\code{(x + px) + b + (y + px) + s}}{(x-anchor) + base length +
(y-anchor) + side length}
\item{\code{(x + px) + b + (y + px) + sb}}{(x-anchor) + base length +
(y-anchor) + base/side angle}
\item{\code{(x + px) + b + (y + px) + ss}}{(x-anchor) + base length +
(y-anchor) + side-side angle}
\item{\code{(x + px) + b + y + y2}}{(x-anchor) + base length + bottom +
top}
\item{\code{x + x2 + (y + py) + h }}{left + right + (y-anchor) + height}
\item{\code{x + x2 + y + y2}}{left + right + bottom + top}}
}

\section{Argument recycling}{
 Arguments \code{x}, \code{y}, \code{loc},
\code{px}, \code{py}, \code{b}, \code{h}, \code{s}, \code{ab}, \code{as},
\code{au}, \code{x2}, and \code{y2} are recycled.

 The argument set \code{\{x, y, b, h, e, re, te,
  be, ee, au, px, py, loc\}} is recycled.
}

\section{The \code{h} and \code{s} arguments}{
 \code{h}) is calculated as the
perpendicular distance from the vertex between non-base sides to the base
side. \code{s} is defined as the length of the two equivalent-length
non-base sides.
}

\section{Defining bounding rectangles}{
 Bounding rectangles can be located in
a number of ways. The following shows valid combination of non-\code{NA}
locating arguments: \describe{
\item{\code{(x + px) + b + (y + py) + h}}{(x-anchor) + base + (y-anchor)
+ height}
\item{\code{(x + px) + b + (y + px) + s}}{(x-anchor) + base + (y-anchor)
+ side length}
\item{\code{(x + px) + b + (y + px) + sb}}{(x-anchor) + base + (y-anchor)
+ base/side angle}
\item{\code{(x + px) + b + (y + px) + ss}}{(x-anchor) + base + (y-anchor)
+ side-side angle}
\item{\code{(x + px) + b + y + te}}{(x-anchor) + base + bottom + top}
\item{\code{x + re + (y + py) + h }}{left + right + (y-anchor) + height}
\item{\code{x + re + y + te}}{left + right + bottom + top}}
}

\seealso{
Other new_shape: 
\code{\link{new_pvtri}()},
\code{\link{new_right}()}

Other new_triangle: 
\code{\link{new_pvtri}()},
\code{\link{new_right}()}

Other new_triangles: 
\code{\link{new_rights}()},
\code{\link{new_triangles}()}
}
\concept{new_shape}
\concept{new_triangle}
\concept{new_triangles}
