% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_pj.R
\name{new_pj}
\alias{new_pj}
\title{Creates a new pj object}
\usage{
new_pj(
  xmin,
  xmax,
  ymin,
  ymax,
  w = 6.5,
  h = 4.5,
  pad = 0.01,
  pu = "in",
  fc = "grey95",
  lc = "black",
  lw = 1
)
}
\arguments{
\item{xmin}{numeric scalar giving the minimum x-value for plotting objects in
the background region.}

\item{xmax}{numeric scalar giving the maximum x-value for plotting objects in
the background region.}

\item{ymin}{numeric scalar giving the minimum y-value for plotting objects in
the background region.}

\item{ymax}{numeric scalar giving the maximum y-value for plotting objects in
the background region.}

\item{w}{positive numeric scalar giving the width of the background in
inches, centimeters, or millimeters.}

\item{h}{positive numeric scalar giving the height of the background in
inches, centimeters, or millimeters.}

\item{pad}{proportion vector of length 1, 2, or 4 (see details section) with
value no larger than 0.25.}

\item{pu}{character scalar defining units in which \code{w} and \code{h} are
expressed. Valid values are \code{'in'}, \code{'cm'}, and \code{'mm'} for
inches, centimeters, and millimeters.}

\item{fc}{character scalar fill color value.}

\item{lc}{character scalar line color value.}

\item{lw}{positive numeric scalar line width (default = 1).}
}
\value{
A pj object (a list of specific tibbles).
}
\description{
Set the minimum and maximum x and y values of a pj object
background its fill color, line color, line width, and edge padding (where
plotting does not occur).
}
\section{Edge padding}{
 Edge padding can be achieved in three ways using the
\code{pad} argument as shown in the following table:
\tabular{llll}{
\emph{edge}  \tab\emph{\code{pad = p}}\tab\emph{\code{pad = c(px, py)}}
\tab\emph{\code{pad = c(pl, pr, pb, pt)}}
\cr\emph{left}  \tab\code{p \* min(w, h)}\tab\code{px \* w}\tab\code{pl \* w}
\cr\emph{right} \tab\code{p \* min(w, h)}\tab\code{px \* w}\tab\code{pr \* w}
\cr\emph{bottom}\tab\code{p \* min(w, h)}\tab\code{py \* h}\tab\code{pb \* h}
\cr\emph{top}   \tab\code{p \* min(w, h)}\tab\code{py \* h}\tab\code{pt \* h}
}
}

