% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_simples.R
\name{new_points}
\alias{new_points}
\title{Add a points shape to a plot}
\usage{
new_points(
  pj,
  x,
  y,
  order = "x",
  region = ".",
  look = NULL,
  mod = NULL,
  name = "."
)
}
\arguments{
\item{pj}{an object of class \code{'pj'}.}

\item{x}{numeric vector defining the horizontal location of each point.}

\item{y}{numeric vector defining the vertical location of each point.}

\item{order}{character scalar \code{'x'}, \code{'y'}, or \code{'i'}
indicating, respectively, ordering x-y pairs on the value of \code{x},
ordering x-y pairs on the value of \code{y}, and identity (leaving the
order of x-y pairs unchanged).}

\item{region}{either character scalar \code{"."} to indicate the most
recently used region, a positive integer to indicate the ID number of a
region (\code{0} indicates the background region), or a character scalar to
indicate the name of a region (\code{'bg'} indicates the background
region).}

\item{look}{an optional list containing up to four lists specifying looks to
apply to the shape. See the \emph{using \code{look} to make shapes appear}
section.}

\item{mod}{an optional list containing up to four lists specifying
modifications to apply to the shape. See the \emph{using \code{mod} to
transform shapes} section.}

\item{name}{character scalar indicating a name for the shape. The special
value \code{"."} indicates that the pj package create a default name for
the shape.}
}
\value{
\code{pj} with the addition of the specified point shape.
}
\description{
A points shape is one or more x-y pairs.
}
\section{Using \code{look} to make shapes appear}{
 By itself, this function
simply defines the location of x-y pairs in the assigned region for the
specified shape. Various looks can be added to the shape using the
\code{look} parameter or via the \code{\link{add_ci}},
\code{\link{add_glyphs}}, \code{\link{add_labels}}, and
\code{\link{add_marks}} functions. Sub-lists of the look argument must be
named \code{'ci'}, \code{'glyph'}, \code{'label'}, and/or \code{'mark'} to
indicate that a confidence interval, glyphs (single characters), labels,
and/or point marks be added to the shape.
}

\section{Argument recycling}{
 The argument set \code{\{x, y, order\}} is
recycled.
}

\section{Using \code{mod} to transform shapes}{
 Shapes can be transformed in
four ways: dilation, reflection, rotation, and translation/. using the
\code{mod} argument or the \code{\link{mod_dilate}},
\code{\link{mod_reflect}}, \code{\link{mod_rotate}}, and
\code{\link{mod_translate}}. Sub-lists of the \code{mod} argument must be
named \code{'dilate'}, \code{'reflect'}, \code{'rotate'}, and/or
\code{'translate'} to indicate that a dilation, reflection, rotation, and
or translation be applied to the shape.
}

\seealso{
Other new_simples: 
\code{\link{new_reflines}()},
\code{\link{new_segments}()}
}
\concept{new_simples}
