% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_quads.R
\name{new_quads}
\alias{new_quads}
\title{Add a proportional-vertex quadrilaterals shape to a plot}
\usage{
new_quads(
  pj,
  pvx,
  pvy,
  x,
  y,
  re = NA,
  te = NA,
  w = NA,
  h = NA,
  px = 0.5,
  py = 0.5,
  region = ".",
  look = NULL,
  mod = NULL,
  name = "."
)
}
\arguments{
\item{pj}{an object of class \code{'pj'}.}

\item{pvx}{proportion vector of length 4 indicating the proportional x
locations of the four vertices of each quad to be plotted as a proportion
of the distance from the left to right edges of their bounding rectangles.
All values must be between 0 and 1, inclusive. See the \emph{using
\code{pvx} and \code{pvy} to locate vertices} section.}

\item{pvy}{proportion vector of length 4 indicating the proportional y
locations of the four vertices of each quad to be plotted as a proportion
of the distance from the bottom to top edges of their bounding rectangles.
All values must be between 0 and 1, inclusive. \emph{using \code{pvx} and
\code{pvy} to locate vertices} section.}

\item{x}{numeric vector giving horizontal location. When \code{x} is used in
conjunction with \code{px} and \code{w} these vectors define a fixed
horizontal point for each shape. When \code{x} is used in conjunction with
\code{re}, \code{x} defines the left edge of the bounding rectangle of each
shape.}

\item{y}{numeric vector giving vertical location. When \code{y} is used in
conjunction with \code{py} and \code{h} these vectors define a fixed
vertical point for each shape. When \code{y} is used in conjunction with
\code{te},\code{y} defines the bottom edge of the bounding rectangle of
each shape.}

\item{re}{\code{NA} or a numeric vector defining the right edge of each
bounding rectangle.}

\item{te}{\code{NA} or a numeric vector defining the top edge of each
bounding rectangle.}

\item{w}{\code{NA} or a numeric vector defining the width of each bounding
rectangle.}

\item{h}{\code{NA} or a numeric vector defining the height of each bounding
rectangle.}

\item{px}{\code{NA} or a proportion vector defining the location of \code{x}
as the proportion of the distance from left to right edge of bounding
rectangles. Values are generally between 0 and 1, inclusive, however, when
they are outside those values, fixed points \code{x} will lie outside
bounding rectangles.}

\item{py}{\code{NA} or a proportion vector defining the location of \code{y}
as the proportion of the distance from bottom to top edge of bounding
rectangles. Values are generally between 0 and 1, inclusive, however, when
they are outside those values, fixed points \code{y} will lie outside
bounding rectangles.}

\item{region}{either character scalar \code{"."} to indicate the most
recently used region, a positive integer to indicate the ID number of a
region (\code{0} indicates the background region), or a character scalar to
indicate the name of a region (\code{'bg'} indicates the background
region).}

\item{look}{an optional list containing up to four lists specifying looks to
apply to the shape. See the \emph{using \code{look} to make shapes appear}
section.}

\item{mod}{an optional list containing up to four lists specifying
modifications to apply to the shape. See the \emph{using \code{mod} to
transform shapes} section.}

\item{name}{character scalar indicating a name for the shape. The special
value \code{"."} indicates that the pj package create a default name for
the shape.}
}
\value{
\code{pj} with the addition of the defined proportional-vertext
quadrilaterals.
}
\description{
A proportional-vertex quadrilaterals shape is a collection of
quadrilaterals defined by the proportional locations of their vertices that
use the same subset of locating parameters from among \code{x}, \code{y}, \code{w}, \code{y},
\code{re}, and \code{te}. Proportional-vertex quadrilaterals are created by
proportionally defining vertex locations within a bounding rectangle. These
shapes are located in plotting regions using bounding rectangles; a concept
used in defining function parameters. See the \emph{locating bounding
rectangles} section for details.
}
\section{Using \code{pvx} and \code{pvy} to locate vertices}{
 \code{pvx}
indicates, as a proportion, how far across the bounding rectangle each of
the four vertices of all quads are located (from left to right). \code{pvy}
indicates, as a proportion, how far up the bounding rectangle each of the
four vertices of all quads are located. For example: \describe{
\item{pvx = c(0, 1/2, 1, 1/2), pvy = c(1/2, 1, 1/2, 0)}{Produces a
\link[=new_diamonds]{diamond/rhombus}, or a kite that is symmetric
both horizontally and vertically.}
\item{pvx = c(0, 0, 1, 1), pvy = c(0, 1, 1, 0)}{Produces a rectangle
identical to the boundary rectangle.}
\item{pvx = c(0, 0.2, 0.8, 1), pvy = c(0, 1, 1, 0)}{Produces a
horizontally-oriented isosceles trapezoid whose top edge is 60\% the
width of its bottom edge.}
\item{pvx = c(0, 0.2, 1, 0.8), pvy = c(0, 1, 1, 0)}{Produces a
horizontally-oriented, right-leaning parallelogram whose top and
bottom edges are 60\% of the width of the bounding rectangle.}
\item{pvx = c(0, 1, 0, 1), pvy = c(0, 1, 1, 0)}{Produces a
vertically-oriented hourglass.}
\item{pvx = c(0, 0, 1, 1), pvy = c(0, 1, 0, 1)}{Produces a
horizontally-oriented hourglass.}
\item{pvx = c(0, 0.8, 1, 0.2), pvy = c(0.2, 1, 0.8, 0)}{Produces a quad
whose first through fourth vertices are, respectively, 20\% up the
left edge of the bounding rectangle, 80\% of the way across the top
edge, 80\% of the way up the right edge, and 20\% of the way across
the bottom edge.}}
}

\section{Defining bounding rectangles}{
 Bounding rectangles can be defined in
a number of ways. The following shows valid combination of non-\code{NA}
locating arguments: \describe{
\item{\code{(x + px) + w + (y + py) + h}}{(horizontal anchor) + width +
(vertical anchor) + height}
\item{\code{(x + px) + w + y + te}}{(horizontal anchor) + bottom edge +
top edge}
\item{\code{x + re + (y + py) + h}}{left edge + right edge + (vertical
anchor) + height}
\item{\code{x + re + y + te}}{left edge + right edge + bottom edge + top
edge}}
}

\section{Bounding rectangles, \code{px}, and \code{py}}{
 When \code{px} is not
\code{NA}, it represents where the value \code{x} is located as a
proportion of the distance between the left and right edge of each bounding
rectangle. Likewise, when \code{py} is not \code{NA}, it represents where
the value \code{y} is located as a proportion of the distance between the
bottom and top edge of each bounding rectangle. For example: \itemize{
\item when \code{px = 0.5} and \code{py = 0.5}, the center of the
bounding rectangle is located at \code{(x, y)}.
\item when \code{px = 0} and \code{py = 0}, the bottom left corner of the
bounding rectangle is located at \code{(x, y)}.
\item when \code{px = 1} and \code{py = 1}, the top right corner of the
bounding rectangle is located at \code{(x, y)}.
\item when \code{px = 1/3} and \code{py = 2/3}, the point 1/3 of the way
from left to right edge and 2/3 of the way from bottom to top edge
of the bounding rectangle is located at \code{(x, y)}.}
\code{px} and \code{py} can be less than 0 or greater than 1, resulting in
the fixed point \code{(x, y)} being outside the bounding rectangle.
}

\section{Using \code{look} to make shapes appear}{
 By itself, this function
simply defines the location of x-y pairs in the assigned region for the
specified shape. Various looks can be added to the shape using the
\code{look} parameter or via the \code{\link{add_ends}},
\code{\link{add_fills}}, \code{\link{add_glyphs}},
\code{\link{add_labels}}, \code{\link{add_lines}}, and
\code{\link{add_marks}} functions. Sub-lists of the look argument must be
named \code{'end'}, \code{'fill'}, \code{'glyph'}, \code{'label'},
\code{'line'}, and/or \code{'mark'} to indicate that line end marks, fill
colors, glyphs (single characters), labels, lines, and/or point marks be
added to the shape.
}

\section{Using \code{mod} to transform shapes}{
 Shapes can be transformed in
four ways: dilation, reflection, rotation, and translation/. using the
\code{mod} argument or the \code{\link{mod_dilate}},
\code{\link{mod_reflect}}, \code{\link{mod_rotate}}, and
\code{\link{mod_translate}}. Sub-lists of the \code{mod} argument must be
named \code{'dilate'}, \code{'reflect'}, \code{'rotate'}, and/or
\code{'translate'} to indicate that a dilation, reflection, rotation, and
or translation be applied to the shape.
}

\section{Argument recycling}{
 The argument set \code{\{x, y, w, h, re, te, px,
  py\}} is recycled.
}

\seealso{
Other new_quads: 
\code{\link{new_diamonds}()},
\code{\link{new_kites}()},
\code{\link{new_rects}()}
}
\concept{new_quads}
