% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_simples.R
\name{new_reflines}
\alias{new_reflines}
\title{Add a reference lines shape to a plot}
\usage{
new_reflines(
  pj,
  x1,
  y1,
  ref = "i",
  x2 = NA,
  y2 = NA,
  s = NA,
  i = NA,
  a = NA,
  au = "r",
  region = ".",
  look = NULL,
  mod = NULL,
  name = "."
)
}
\arguments{
\item{x1}{\code{NA} or a numeric vector indicating either a vertical
reference line at the value contained in \code{x1} (if used alone) or the
horizontal location of a primary point used to define a reference line (if
used with \code{y1}).}

\item{y1}{\code{NA} or a numeric vector indicating either a horizontal
reference line at the value contained in\code{y1} (if used alone) or the
vertical location of a primary point used to define a reference line (if
used with \code{x1}).}

\item{ref}{\code{NA} or a character vector giving pre-defined reference lines
where \code{'i'} indicates the identity line \code{'x'} indicates the
x-axis, and \code{'y'} indicates the y-axis.}

\item{x2}{\code{NA} or a numeric vector indicating the horizontal locations
of secondary points to define lines in conjunction with \code{x1},
\code{y1}, and \code{y2}.}

\item{y2}{\code{NA} or a numeric vector indicating the vertical locations of
secondary points to define lines in conjunction with \code{x1}, \code{y1},
and \code{x2}.}

\item{s}{\code{NA} or a numeric vector indicating the slope of each reference
line.}

\item{i}{\code{NA} or a numeric vector indicating the x-intercept of each
reference line.}

\item{a}{\code{NA} or a numeric vector indicating the angle of each reference
line.}

\item{au}{\code{NA} or a character vector indicating the angle units to be
used where \code{'d'} indicates degrees, \code{'g'} indicates gradians,
\code{'p'} indicates proportion of a full revolution, and \code{'r'}
indicates radians.}
}
\value{
\code{pj} with the addition of the specified refline shape(s).
}
\description{
A reference lines shape is a collection of lines going from edge
to edge of a plotting region (without regard to any pad reserved around the
edges of the region) that use the same subset of locating arguments
(\code{x1}, \code{y1}, \code{a}, \code{i}, \code{s}, \code{x2}, \code{y2},
and \code{au}). Reference lines can be defined in a variety of ways as
described in the \emph{defining reference lines} section.
}
\section{Using \code{look} to make shapes appear}{
 By itself, this function
simply defines the location of x-y pairs in the assigned region for the
specified shape. Various looks can be added to the shape using the
\code{look} parameter or via the \code{\link{add_lines}} function. The
sub-list of the look argument must be named \code{'line'}.
}

\section{Argument recycling}{
 The argument set \code{\{ref, x1, y1, x2, y2, s,
  i, a, au\}} is recycled.
}

\section{Defining reference lines}{
 Valid combinations of non-\code{NA}
locating arguments are given below. \describe{
\item{\code{ref}}{Pre-defined reference line}
\item{\code{x1}}{Vertical line at \code{x1}.}
\item{\code{y1}}{Horizontal line at \code{y1}.}
\item{\code{x1 + y1 + x2 + y2}}{Point-point definition (line through
\code{(x1, y1)} and \code{(x2, y2)}).}
\item{\code{x1 + y1 + a}}{Point-angle definition (line through
\code{(x1, y1)} at angle \code{a}).}
\item{\code{x1 + y1 + i}}{Point-intercept definition (line through
intercept and \code{(x1, y1)}).}
\item{\code{x1 + y1 + s}}{Point-slope definition (line through
\code{(x1, y1)} with slope \code{s}).}
\item{\code{a + i}}{Angle-intercept definition (line through intercept
at angle \code{a}).}
\item{\code{s + i}}{Slope-intercept definition (line through intercept
with slope \code{s}.}}
}

\seealso{
Other new_simples: 
\code{\link{new_points}()},
\code{\link{new_segments}()}
}
\concept{new_simples}
