% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_region.R
\name{new_region}
\alias{new_region}
\title{Create a new plotting region contained in the background region}
\usage{
new_region(
  pj,
  xmax,
  xmin,
  ymax,
  ymin,
  le = NA,
  be = NA,
  re = NA,
  te = NA,
  w = NA,
  h = NA,
  fc = "white",
  fa = 0.5,
  lc = "black",
  la = 0.5,
  lt = "-",
  lw = 1,
  name = "."
)
}
\arguments{
\item{pj}{an object of class 'pj'.}

\item{xmax}{a numeric scalar giving the maximum x-value for plotting objects
in the region.}

\item{xmin}{a numeric scalar giving the minimum x-value for plotting objects
in the region.}

\item{ymax}{a numeric scalar giving the maximum y-value for plotting objects
in the region.}

\item{ymin}{a numeric scalar giving the minimum y-value for plotting objects
in the region.}

\item{le}{\code{NA} or a numeric scalar locating the left edge of the region
either in x units of the background region or as a proportion of the
distance from the left to right edge of the area of the background region
available for plotting.}

\item{be}{\code{NA} or a numeric scalar locating the bottom edge of the
region either in y units of the background region or as a proportion of the
distance from the bottom to top edge of the area of the background region
available for plotting.}

\item{re}{\code{NA} or a numeric scalar locating the right edge of the region
either in x units of the background region or as a proportion of the
distance from the left to right edge of the area of the background region
available for plotting.}

\item{te}{\code{NA} or a numeric scalar locating the top edge of the region
either in y units of the background region or as a proportion of the
distance from the bottom to top edge of the area of the background region
available for plotting.}

\item{w}{\code{NA} or a positive numeric scalars giving the width of the
region either in x-units of the background region or as a proportion of the
width of the background region that is available for plotting.}

\item{h}{\code{NA} or a positive numeric scalars giving the height of the
region either in x-units of the background region or as a proportion of the
height of the background region that is available for plotting.}

\item{fc}{\code{NA} or a character scalar region fill color value.}

\item{fa}{\code{NA} or a proportion scalars (i.e., between 0 and 1,
inclusive) indicating the alpha opacity level of the region fill color.}

\item{lc}{\code{NA} or a character scalar region line color value.}

\item{la}{\code{NA} or a proportion scalars (i.e., between 0 and 1,
inclusive) indicating the alpha opacity level of the region line color.}

\item{lt}{\code{NA} or a character scalar giving line type. Valid values are
\code{c("-", ":", "=", ":=", "L=", "2=")}. See \code{the lt argument}
section.}

\item{lw}{\code{NA} or a positive numeric scalar line width indicating a
multiple of the default line width (i.e., default = 1).}

\item{name}{character scalar name for the new region. The special value
\code{'.'} indicates that the pj package should create a name for the new
region.}

\item{pad}{proportion vector of length 1, 2, or 4 (see the \emph{edge
padding} section) with value no larger than 0.25.}
}
\value{
\code{pj} with the addition of the specified region.
}
\description{
Locate a new plotting region within the background region, set
its units relative to the background region, set its min and max x and y
values, put a pad between the edges of the region and available area for
plotting, and apply fill and line looks.
}
\section{The lt argument}{
 The \code{lt} argument can take on the following
values to indicate the following line types:\tabular{ll}{
VALUE      \tab LINE TYPE          \cr
\code{'-'} \tab solid              \cr
\code{':'} \tab dotted             \cr
\code{'='} \tab dashed             \cr
\code{':='}\tab dot-dashed         \cr
\code{'L='}\tab long-dashed        \cr
\code{'2='}\tab double-width dashed}
}

\section{Edge padding}{
 Edge padding can be achieved in three ways using the
\code{pad} argument as shown in the following table:
\tabular{llll}{
\emph{edge}  \tab\emph{\code{pad = p}}\tab\emph{\code{pad = c(px, py)}}
\tab\emph{\code{pad = c(pl, pr, pb, pt)}}
\cr\emph{left}  \tab\code{p \* min(w, h)}\tab\code{px \* w}\tab\code{pl \* w}
\cr\emph{right} \tab\code{p \* min(w, h)}\tab\code{px \* w}\tab\code{pr \* w}
\cr\emph{bottom}\tab\code{p \* min(w, h)}\tab\code{py \* h}\tab\code{pb \* h}
\cr\emph{top}   \tab\code{p \* min(w, h)}\tab\code{py \* h}\tab\code{pt \* h}
}
}

