% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_polys.R
\name{new_regulars}
\alias{new_regulars}
\title{Add a regular polygons shape to a plot}
\usage{
new_regulars(
  pj,
  x,
  y,
  n,
  r = NA,
  e = NA,
  a = NA,
  px = 0.5,
  py = 0.5,
  loc = "b",
  region = ".",
  look = NULL,
  mod = NULL,
  name = "."
)
}
\arguments{
\item{pj}{an object of class \code{'pj'}.}

\item{x}{numeric vector giving horizontal location. When \code{x} is used in
conjunction with \code{px} and \code{w} these vectors define a fixed
horizontal point for each shape. When \code{x} is used in conjunction with
\code{re}, \code{x} defines the left edge of the bounding rectangle of each
shape.}

\item{y}{numeric vector giving vertical location. When \code{y} is used in
conjunction with \code{py} and \code{h} these vectors define a fixed
vertical point for each shape. When \code{y} is used in conjunction with
\code{te},\code{y} defines the bottom edge of the bounding rectangle of
each shape.}

\item{n}{integer vector with values between 3 and 1000, inclusive, indicating
the numbers of sides/vertices of each regular polygon.}

\item{e}{\code{NA} or a positive numeric vector indicating edge length of
each polygon.}

\item{px}{\code{NA} or a proportion vector defining the location of \code{x}
as the proportion of the distance from left to right edge of bounding
rectangles. Values are generally between 0 and 1, inclusive, however, when
they are outside those values, fixed points \code{x} will lie outside
bounding rectangles.}

\item{py}{\code{NA} or a proportion vector defining the location of \code{y}
as the proportion of the distance from bottom to top edge of bounding
rectangles. Values are generally between 0 and 1, inclusive, however, when
they are outside those values, fixed points \code{y} will lie outside
bounding rectangles.}

\item{loc}{character vector indicating where the base of each regular polygon
is located. \code{'l'}, \code{'r'}, \code{'b'}, and \code{'t'} indicate
left, right, bottom, and top edges of the bounding rectangle, respectively.}

\item{region}{either character scalar \code{"."} to indicate the most
recently used region, a positive integer to indicate the ID number of a
region (\code{0} indicates the background region), or a character scalar to
indicate the name of a region (\code{'bg'} indicates the background
region).}

\item{look}{an optional list containing up to four lists specifying looks to
apply to the shape. See the \emph{using \code{look} to make shapes appear}
section.}

\item{mod}{an optional list containing up to four lists specifying
modifications to apply to the shape. See the \emph{using \code{mod} to
transform shapes} section.}

\item{name}{character scalar indicating a name for the shape. The special
value \code{"."} indicates that the pj package create a default name for
the shape.}
}
\value{
\code{pj} with the addition of the defined regular polygons.
}
\description{
A regular polygons shape is a collection of regular polygons
using the ame subset of locating arguments from among \code{x}, \code{y},
\code{n}, \code{loc}, \code{px}, \code{py}, \code{s}, \code{x2}, and
\code{y2}. Regular polygons are located in plotting regions using bounding
rectangles; a concept used in defining function parameters. See the
\emph{defining bounding rectangles} section for details.
}
\section{Defining bounding rectangles}{
 Bounding rectangles can be located in
a number of ways. The following shows valid combination of non-\code{NA}
locating arguments for regular polygons: \describe{
\item{\code{(x + px) + (y + py) + e}}{(horizontal anchor) + (vertical
anchor) + edge length}
\item{\code{(x + px) + y + te}}{(horizontal anchor) + bottom edge + top
edge}
\item{\code{x + re + (y + px)}}{left edge + right edge + (vertical
anchor)}}
In addition, regular polygons can be squeezed or stretched horizontally and
vertically by specifying the location of all four edges of the bounding
rectangle as follows: \describe{
\item{\code{x + re + y + te)}}{left edge + right edge +  bottom edge +
top edge}}
}

\section{Argument recycling}{
 The argument set \code{\{n, x, y, e, re, te, px,
  py, loc\}} is recycled.
}

\section{Bounding rectangles, \code{px}, and \code{py}}{
 When \code{px} is not
\code{NA}, it represents where the value \code{x} is located as a
proportion of the distance between the left and right edge of each bounding
rectangle. Likewise, when \code{py} is not \code{NA}, it represents where
the value \code{y} is located as a proportion of the distance between the
bottom and top edge of each bounding rectangle. For example: \itemize{
\item when \code{px = 0.5} and \code{py = 0.5}, the center of the
bounding rectangle is located at \code{(x, y)}.
\item when \code{px = 0} and \code{py = 0}, the bottom left corner of the
bounding rectangle is located at \code{(x, y)}.
\item when \code{px = 1} and \code{py = 1}, the top right corner of the
bounding rectangle is located at \code{(x, y)}.
\item when \code{px = 1/3} and \code{py = 2/3}, the point 1/3 of the way
from left to right edge and 2/3 of the way from bottom to top edge
of the bounding rectangle is located at \code{(x, y)}.}
\code{px} and \code{py} can be less than 0 or greater than 1, resulting in
the fixed point \code{(x, y)} being outside the bounding rectangle.
}

\section{Using \code{look} to make shapes appear}{
 By itself, this function
simply defines the location of x-y pairs in the assigned region for the
specified shape. Various looks can be added to the shape using the
\code{look} parameter or via the \code{\link{add_ends}},
\code{\link{add_fills}}, \code{\link{add_glyphs}},
\code{\link{add_labels}}, \code{\link{add_lines}}, and
\code{\link{add_marks}} functions. Sub-lists of the look argument must be
named \code{'end'}, \code{'fill'}, \code{'glyph'}, \code{'label'},
\code{'line'}, and/or \code{'mark'} to indicate that line end marks, fill
colors, glyphs (single characters), labels, lines, and/or point marks be
added to the shape.
}

\section{Using \code{mod} to transform shapes}{
 Shapes can be transformed in
four ways: dilation, reflection, rotation, and translation/. using the
\code{mod} argument or the \code{\link{mod_dilate}},
\code{\link{mod_reflect}}, \code{\link{mod_rotate}}, and
\code{\link{mod_translate}}. Sub-lists of the \code{mod} argument must be
named \code{'dilate'}, \code{'reflect'}, \code{'rotate'}, and/or
\code{'translate'} to indicate that a dilation, reflection, rotation, and
or translation be applied to the shape.
}

\seealso{
Other new_polygons: 
\code{\link{new_free}()},
\code{\link{new_polys}()},
\code{\link{new_stars}()}
}
\concept{new_polygons}
