% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_polycurve.R
\name{new_ribbon}
\alias{new_ribbon}
\title{Add a ribbon shape to a plot}
\usage{
new_ribbon(
  pj,
  order = "x",
  dir = "v",
  x1 = NA,
  y1 = NA,
  x2 = NA,
  y2 = NA,
  region = ".",
  look = NULL,
  mod = NULL,
  name = "."
)
}
\arguments{
\item{pj}{an object of class \code{'pj'}.}

\item{order}{character scalar \code{'x'}, \code{'y'}, or \code{'i'}
indicating ordering x-y pairs on the value of \code{x} and \code{x2},
ordering x-y pairs on the value of \code{y} and \code{y2}, and identity, or
leaving the order of x-y pairs unchanged.}

\item{dir}{character scalar identifying the directional orientation of the
ribbon. \code{'v'} indicates a vertically-oriented ribbon, or the vertical
area between the two curves. \code{'h'} indicates a horizontally-oriented
ribbon.}

\item{x1}{numeric vector giving the horizontal location of points on the
first curve. The longer of \code{x} or \code{y} must have at least 2
values.}

\item{y1}{numeric vector giving the vertical location of points on the first
curve. The longer of \code{x} or \code{y} must have at least 2 values.}

\item{x2}{numeric vector giving the horizontal location of points on the
second curve. The longer of \code{x2} or \code{y2} must have at least 3
values.}

\item{y2}{numeric vector giving the vertical location of points on the second
curve. The longer of \code{x2} or \code{y2} must have at least 3 values.}

\item{region}{either character scalar \code{"."} to indicate the most
recently used region, a positive integer to indicate the ID number of a
region (\code{0} indicates the background region), or a character scalar to
indicate the name of a region (\code{'bg'} indicates the background
region).}

\item{look}{an optional list containing up to four lists specifying looks to
apply to the shape. See the \emph{using \code{look} to make shapes appear}
section.}

\item{mod}{an optional list containing up to four lists specifying
modifications to apply to the shape. See the \emph{using \code{mod} to
transform shapes} section.}

\item{name}{character scalar indicating a name for the shape. The special
value \code{"."} indicates that the pj package create a default name for
the shape.}
}
\value{
\code{pj} with the addition of the defined ribbon.
}
\description{
A ribbon shape is a polygon defined as the area between
two curves (two sets of x-y pairs).
}
\section{Argument recycling}{
 The argument set \code{\{x1, y1\}} is recycled.
The argument set \code{\{x2, y2}\} is separately recycled.
}

\section{Using \code{look} to make shapes appear}{
 By itself, this function
simply defines the location of x-y pairs in the assigned region for the
specified shape. Various looks can be added to the shape using the
\code{look} parameter or via the \code{\link{add_ends}},
\code{\link{add_fills}}, \code{\link{add_glyphs}},
\code{\link{add_labels}}, \code{\link{add_lines}}, and
\code{\link{add_marks}} functions. Sub-lists of the look argument must be
named \code{'end'}, \code{'fill'}, \code{'glyph'}, \code{'label'},
\code{'line'}, and/or \code{'mark'} to indicate that line end marks, fill
colors, glyphs (single characters), labels, lines, and/or point marks be
added to the shape.
}

\section{Using \code{mod} to transform shapes}{
 Shapes can be transformed in
four ways: dilation, reflection, rotation, and translation/. using the
\code{mod} argument or the \code{\link{mod_dilate}},
\code{\link{mod_reflect}}, \code{\link{mod_rotate}}, and
\code{\link{mod_translate}}. Sub-lists of the \code{mod} argument must be
named \code{'dilate'}, \code{'reflect'}, \code{'rotate'}, and/or
\code{'translate'} to indicate that a dilation, reflection, rotation, and
or translation be applied to the shape.
}

\seealso{
Other new_polycurve: 
\code{\link{new_area}()}
}
\concept{new_polycurve}
