% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_triangle.R
\name{new_right}
\alias{new_right}
\title{Add one or more right triangle shapes}
\usage{
new_right(
  jp,
  x,
  y,
  loc = "bl",
  px = 0.5,
  py = 0.5,
  b = NA,
  h = NA,
  hyp = NA,
  hb = NA,
  hh = NA,
  x2 = NA,
  y2 = NA,
  au = NA,
  look = NULL,
  mod = NULL,
  region = ".",
  name = "."
)
}
\arguments{
\item{loc}{(required) character vector indicating where the 90-degree angle
of the right triangles are located. \code{'bl'}, \code{'br'}, \code{'tl'},
and \code{'tr'} indicate bottom-left, bottom-right, top-left, and top-right
corners of the bounding rectangle, respectively.}

\item{b}{(optional) positive numeric vector indicating base length.}

\item{hyp}{(optional) positive numeric vector indicating hypotenuse length.}

\item{hb, hh}{(optional) positive numeric vectors indicating, respectively,
the angle between (\code{hb}) the hypotenuse and base and (\code{hs}) the
angle between the hypotenuse and the non-base side.}

\item{au}{(optional) character vector indicating angle units where
\code{'d'}, \code{'g'}, \code{'r'}, and \code{'p'} indicate degrees,
gradians, radians, and proportions of a revolution.}
}
\value{
\code{jp} with the addition of the specified right triangle shape(s).
}
\description{
right triangle shapes are located in plotting regions using
bounding rectangles; a concept used in defining function parameters. See
the \emph{locating bounding rectangles} section for details.
}
\section{Locating bounding rectangles}{
 Bounding rectangles can be located in
a number of ways. The following shows valid combination of non-\code{NA}
locating arguments:
\describe{
\item{\code{(x + px) + (y + py) + b + h}}{(x anchor) + (y anchor) + base
+ height}
\item{\code{(x + px) + (y + py) + b + hyp}}{(x anchor) + (y anchor) +
base + hypotenuse}
\item{\code{(x + px) + (y + py) + h + hyp}}{(x anchor) + (y anchor) +
height + hypotenuse}
\item{\code{(x + px) + (y + py) + b + hb}}{(x anchor) + (y anchor) +
base + base/hypotenuse angle}
\item{\code{(x + px) + (y + py) + b + hh}}{(x anchor) + (y anchor) +
base + height/hypotenuse angle}
\item{\code{(x + px) + (y + py) + h + hb}}{(x anchor) + (y anchor) +
height + base/hypotenuse angle}
\item{\code{(x + px) + (y + py) + h + hh}}{(x anchor) + (y anchor) +
height + height/hypotenuse angle}
\item{\code{(x + px) + (y + py) + hyp + hb}}{(x anchor) + (y anchor) +
hypotenuse + base/hypotenuse angle}
\item{\code{(x + px) + (y + py) + hyp + hh}}{(x anchor) + (y anchor) +
hypotenuse + height/hypotenuse angle}
\item{\code{(x + px) + b + y + y2}}{(x-anchor) + base + bottom edge + top
edge}
\item{\code{x + x2 + (y + py) + h}}{left edge + right edge + (y-anchor) +
height}
\item{\code{x + x2 + y + y2}}{left edge + right edge + bottom edge + top
edge}}
}

\section{Argument recycling}{
 Arguments \code{x}, \code{y}, \code{loc},
\code{px}, \code{py}, \code{b}, \code{h}, \code{s}, \code{ab}, \code{as},
\code{au}, \code{x2}, and \code{y2} are recycled.
}

\seealso{
Other new_shape: 
\code{\link{new_isosceles}()},
\code{\link{new_pvtri}()}

Other new_triangle: 
\code{\link{new_isosceles}()},
\code{\link{new_pvtri}()}
}
\concept{new_shape}
\concept{new_triangle}
