% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_triangles.R
\name{new_rights}
\alias{new_rights}
\title{Add a right triangles shape to a plot}
\usage{
new_rights(
  pj,
  x,
  y,
  b = NA,
  h = NA,
  hyp = NA,
  re = NA,
  te = NA,
  hb = NA,
  hh = NA,
  loc = "bl",
  px = 0.5,
  py = 0.5,
  au = "r",
  region = ".",
  look = NULL,
  mod = NULL,
  name = "."
)
}
\arguments{
\item{pj}{an object of class \code{'pj'}.}

\item{x}{numeric vector giving horizontal location. When \code{x} is used in
conjunction with \code{px} and \code{w} these vectors define a fixed
horizontal point for each shape. When \code{x} is used in conjunction with
\code{re}, \code{x} defines the left edge of the bounding rectangle of each
shape.}

\item{y}{numeric vector giving vertical location. When \code{y} is used in
conjunction with \code{py} and \code{h} these vectors define a fixed
vertical point for each shape. When \code{y} is used in conjunction with
\code{te},\code{y} defines the bottom edge of the bounding rectangle of
each shape.}

\item{b}{\code{NA} or a positive numeric vector indicating base length.}

\item{h}{\code{NA} or a numeric vector defining the height of each bounding
rectangle.}

\item{hyp}{\code{NA} or a positive numeric vector indicating hypotenuse
length.}

\item{re}{\code{NA} or a numeric vector defining the right edge of each
bounding rectangle.}

\item{te}{\code{NA} or a numeric vector defining the top edge of each
bounding rectangle.}

\item{hb}{\code{NA} or a positive numeric vectors indicating, the angle
between the hypotenuse and base sides.}

\item{hh}{\code{NA} or a positive numeric vectors indicating, respectively,
the angle between the hypotenuse and height sides.}

\item{loc}{character vector indicating where the 90-degree angle of the right
triangles are located. \code{'bl'}, \code{'br'}, \code{'tl'}, and
\code{'tr'} indicate bottom-left, bottom-right, top-left, and top-right
corners of the bounding rectangle, respectively.}

\item{px}{\code{NA} or a proportion vector defining the location of \code{x}
as the proportion of the distance from left to right edge of bounding
rectangles. Values are generally between 0 and 1, inclusive, however, when
they are outside those values, fixed points \code{x} will lie outside
bounding rectangles.}

\item{py}{\code{NA} or a proportion vector defining the location of \code{y}
as the proportion of the distance from bottom to top edge of bounding
rectangles. Values are generally between 0 and 1, inclusive, however, when
they are outside those values, fixed points \code{y} will lie outside
bounding rectangles.}

\item{au}{\code{NA} or a character vector indicating angle units where
\code{'d'}, \code{'g'}, \code{'r'}, and \code{'p'} indicate degrees,
gradians, radians, and proportions of a revolution.}

\item{region}{either character scalar \code{"."} to indicate the most
recently used region, a positive integer to indicate the ID number of a
region (\code{0} indicates the background region), or a character scalar to
indicate the name of a region (\code{'bg'} indicates the background
region).}

\item{look}{an optional list containing up to four lists specifying looks to
apply to the shape. See the \emph{using \code{look} to make shapes appear}
section.}

\item{mod}{an optional list containing up to four lists specifying
modifications to apply to the shape. See the \emph{using \code{mod} to
transform shapes} section.}

\item{name}{character scalar indicating a name for the shape. The special
value \code{"."} indicates that the pj package create a default name for
the shape.}
}
\value{
\code{pj} with the addition of the defined right triangles.
}
\description{
Right triangle shapes defined as a collection of right triangles
sharing the same subset of locating arguments (\code{x}, \code{y},
\code{loc}, \code{px}, \code{py}, \code{b}, \code{h}, \code{e}, \code{re},
and \code{te}, \code{hb}, \code{hh}, \code{au}). Right triangles are
located in plotting regions using bounding rectangles, a concept used in
defining function parameters. See the \emph{defining bounding rectangles}
section for details.
}
\section{Defining bounding rectangles}{
 Bounding rectangles can be located in
a number of ways. The following shows valid combination of non-\code{NA}
locating arguments:
\describe{
\item{\code{(x + px) + (y + py) + b + h}}{(x anchor) + (y anchor) + base
+ height}
\item{\code{(x + px) + (y + py) + b + hyp}}{(x anchor) + (y anchor) +
base + hypotenuse}
\item{\code{(x + px) + (y + py) + h + hyp}}{(x anchor) + (y anchor) +
height + hypotenuse}
\item{\code{(x + px) + (y + py) + b + hb}}{(x anchor) + (y anchor) +
base + base/hypotenuse angle}
\item{\code{(x + px) + (y + py) + b + hh}}{(x anchor) + (y anchor) +
base + height/hypotenuse angle}
\item{\code{(x + px) + (y + py) + h + hb}}{(x anchor) + (y anchor) +
height + base/hypotenuse angle}
\item{\code{(x + px) + (y + py) + h + hh}}{(x anchor) + (y anchor) +
height + height/hypotenuse angle}
\item{\code{(x + px) + (y + py) + hyp + hb}}{(x anchor) + (y anchor) +
hypotenuse + base/hypotenuse angle}
\item{\code{(x + px) + (y + py) + hyp + hh}}{(x anchor) + (y anchor) +
hypotenuse + height/hypotenuse angle}
\item{\code{(x + px) + b + y + te}}{(x-anchor) + base + bottom edge + top
edge}
\item{\code{x + re + (y + py) + h}}{left edge + right edge + (y-anchor) +
height}
\item{\code{x + re + y + te}}{left edge + right edge + bottom edge + top
edge}}
}

\section{Argument recycling}{
 The argument set \code{\{x, y, b, h, e, re, te,
  be, ee, au, px, py, loc\}} is recycled.
}

\section{Bounding rectangles, \code{px}, and \code{py}}{
 When \code{px} is not
\code{NA}, it represents where the value \code{x} is located as a
proportion of the distance between the left and right edge of each bounding
rectangle. Likewise, when \code{py} is not \code{NA}, it represents where
the value \code{y} is located as a proportion of the distance between the
bottom and top edge of each bounding rectangle. For example: \itemize{
\item when \code{px = 0.5} and \code{py = 0.5}, the center of the
bounding rectangle is located at \code{(x, y)}.
\item when \code{px = 0} and \code{py = 0}, the bottom left corner of the
bounding rectangle is located at \code{(x, y)}.
\item when \code{px = 1} and \code{py = 1}, the top right corner of the
bounding rectangle is located at \code{(x, y)}.
\item when \code{px = 1/3} and \code{py = 2/3}, the point 1/3 of the way
from left to right edge and 2/3 of the way from bottom to top edge
of the bounding rectangle is located at \code{(x, y)}.}
\code{px} and \code{py} can be less than 0 or greater than 1, resulting in
the fixed point \code{(x, y)} being outside the bounding rectangle.
}

\section{Using \code{look} to make shapes appear}{
 By itself, this function
simply defines the location of x-y pairs in the assigned region for the
specified shape. Various looks can be added to the shape using the
\code{look} parameter or via the \code{\link{add_ends}},
\code{\link{add_fills}}, \code{\link{add_glyphs}},
\code{\link{add_labels}}, \code{\link{add_lines}}, and
\code{\link{add_marks}} functions. Sub-lists of the look argument must be
named \code{'end'}, \code{'fill'}, \code{'glyph'}, \code{'label'},
\code{'line'}, and/or \code{'mark'} to indicate that line end marks, fill
colors, glyphs (single characters), labels, lines, and/or point marks be
added to the shape.
}

\section{Using \code{mod} to transform shapes}{
 Shapes can be transformed in
four ways: dilation, reflection, rotation, and translation/. using the
\code{mod} argument or the \code{\link{mod_dilate}},
\code{\link{mod_reflect}}, \code{\link{mod_rotate}}, and
\code{\link{mod_translate}}. Sub-lists of the \code{mod} argument must be
named \code{'dilate'}, \code{'reflect'}, \code{'rotate'}, and/or
\code{'translate'} to indicate that a dilation, reflection, rotation, and
or translation be applied to the shape.
}

\seealso{
Other new_triangles: 
\code{\link{new_isosceles}()},
\code{\link{new_triangles}()}
}
\concept{new_triangles}
