% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_rounds.R
\name{new_rounds}
\alias{new_rounds}
\title{Add a proportional-vertex rounds shape to a plot}
\usage{
new_rounds(
  pj,
  pv,
  x,
  y,
  w = NA,
  h = NA,
  re = NA,
  te = NA,
  px = 0.5,
  py = 0.5,
  n = 100,
  region = ".",
  look = NULL,
  mod = NULL,
  name = "."
)
}
\arguments{
\item{pj}{an object of class \code{'pj'}.}

\item{pv}{proportion vector of length 4 giving, in order, the vertical
location of the left-edge vertex as a proportion of the distance from
bottom edge to top edge, the bottom-edge vertex as a proportion of the
distance from left edge to right edge, the right-edge vertex as a
proportion of the distance from bottom to top edge, and the top-edge vertex
as a proportion of the distance from left edge to right edge.}

\item{x}{numeric vector giving horizontal location. When \code{x} is used in
conjunction with \code{px} and \code{w} these vectors define a fixed
horizontal point for each shape. When \code{x} is used in conjunction with
\code{re}, \code{x} defines the left edge of the bounding rectangle of each
shape.}

\item{y}{numeric vector giving vertical location. When \code{y} is used in
conjunction with \code{py} and \code{h} these vectors define a fixed
vertical point for each shape. When \code{y} is used in conjunction with
\code{te},\code{y} defines the bottom edge of the bounding rectangle of
each shape.}

\item{w}{\code{NA} or a numeric vector defining the width of each bounding
rectangle.}

\item{h}{\code{NA} or a numeric vector defining the height of each bounding
rectangle.}

\item{re}{\code{NA} or a numeric vector defining the right edge of each
bounding rectangle.}

\item{te}{\code{NA} or a numeric vector defining the top edge of each
bounding rectangle.}

\item{px}{\code{NA} or a proportion vector defining the location of \code{x}
as the proportion of the distance from left to right edge of bounding
rectangles. Values are generally between 0 and 1, inclusive, however, when
they are outside those values, fixed points \code{x} will lie outside
bounding rectangles.}

\item{py}{\code{NA} or a proportion vector defining the location of \code{y}
as the proportion of the distance from bottom to top edge of bounding
rectangles. Values are generally between 0 and 1, inclusive, however, when
they are outside those values, fixed points \code{y} will lie outside
bounding rectangles.}

\item{n}{positive whole-number vector giving the number of vertices to
calculate for plotting each shape. All values must be between 3 and 1000,
inclusive.}

\item{region}{either character scalar \code{"."} to indicate the most
recently used region, a positive integer to indicate the ID number of a
region (\code{0} indicates the background region), or a character scalar to
indicate the name of a region (\code{'bg'} indicates the background
region).}

\item{look}{an optional list containing up to four lists specifying looks to
apply to the shape. See the \emph{using \code{look} to make shapes appear}
section.}

\item{mod}{an optional list containing up to four lists specifying
modifications to apply to the shape. See the \emph{using \code{mod} to
transform shapes} section.}

\item{name}{character scalar indicating a name for the shape. The special
value \code{"."} indicates that the pj package create a default name for
the shape.}
}
\value{
\code{pj} with the addition of the defined proportional-vertex
rounds.
}
\description{
A proportional-vertex rounds shape is a collection of rounded
shapes defined by proportional location of bottom, top, left, and right
vertex locations that use the same subset of locating arguments from among
\code{x}, \code{y}, \code{n}, \code{px}, \code{py}, \code{rx}, \code{ry},
\code{re}, \code{te}, and \code{pv}. Pvrounds are located in plotting
regions using bounding rectangles, a concept used in defining function
parameters. See the \emph{defining bounding rectangles} section for
details. They are defined by specifying, proportionally, using \code{pv},
where from left to right edge the bottom-most and top-most vertices lie,
and from bottom to top edge where left-most and right-most vertices lie.
}
\section{Defining bounding rectangles}{
 Bounding rectangles can be located in
a number of ways. The following shows valid combination of non-\code{NA}
locating arguments: \describe{
\item{\code{(x + px) + w + (y + py) + h}}{(horizontal anchor) + width +
(vertical anchor) + height}
\item{\code{(x + px) + w + y + te}}{(horizontal anchor) + height + bottom
edge + top edge}
\item{\code{x + re + (y + py) + h}}{left edge + right edge + (vertical
anchor) + height}
\item{\code{x + re + y + te}}{left edge + right edge + bottom edge + top
edge}}
}

\section{Argument recycling}{
 The argument set \code{\{x, y, w, h, re, te, px,
  py, n\}} is recycled.
}

\section{Bounding rectangles, \code{px}, and \code{py}}{
 When \code{px} is not
\code{NA}, it represents where the value \code{x} is located as a
proportion of the distance between the left and right edge of each bounding
rectangle. Likewise, when \code{py} is not \code{NA}, it represents where
the value \code{y} is located as a proportion of the distance between the
bottom and top edge of each bounding rectangle. For example: \itemize{
\item when \code{px = 0.5} and \code{py = 0.5}, the center of the
bounding rectangle is located at \code{(x, y)}.
\item when \code{px = 0} and \code{py = 0}, the bottom left corner of the
bounding rectangle is located at \code{(x, y)}.
\item when \code{px = 1} and \code{py = 1}, the top right corner of the
bounding rectangle is located at \code{(x, y)}.
\item when \code{px = 1/3} and \code{py = 2/3}, the point 1/3 of the way
from left to right edge and 2/3 of the way from bottom to top edge
of the bounding rectangle is located at \code{(x, y)}.}
\code{px} and \code{py} can be less than 0 or greater than 1, resulting in
the fixed point \code{(x, y)} being outside the bounding rectangle.
}

\section{Using \code{look} to make shapes appear}{
 By itself, this function
simply defines the location of x-y pairs in the assigned region for the
specified shape. Various looks can be added to the shape using the
\code{look} parameter or via the \code{\link{add_ends}},
\code{\link{add_fills}}, \code{\link{add_glyphs}},
\code{\link{add_labels}}, \code{\link{add_lines}}, and
\code{\link{add_marks}} functions. Sub-lists of the look argument must be
named \code{'end'}, \code{'fill'}, \code{'glyph'}, \code{'label'},
\code{'line'}, and/or \code{'mark'} to indicate that line end marks, fill
colors, glyphs (single characters), labels, lines, and/or point marks be
added to the shape.
}

\section{Using \code{mod} to transform shapes}{
 Shapes can be transformed in
four ways: dilation, reflection, rotation, and translation/. using the
\code{mod} argument or the \code{\link{mod_dilate}},
\code{\link{mod_reflect}}, \code{\link{mod_rotate}}, and
\code{\link{mod_translate}}. Sub-lists of the \code{mod} argument must be
named \code{'dilate'}, \code{'reflect'}, \code{'rotate'}, and/or
\code{'translate'} to indicate that a dilation, reflection, rotation, and
or translation be applied to the shape.
}

\seealso{
Other new_rounds: 
\code{\link{new_circles}()},
\code{\link{new_ellipses}()}
}
\concept{new_rounds}
