% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_simples.R
\name{new_segments}
\alias{new_segments}
\title{Add a segments shape to a plot}
\usage{
new_segments(
  pj,
  x1,
  y1,
  x2 = NA,
  y2 = NA,
  d = NA,
  s = NA,
  a = NA,
  au = "r",
  region = ".",
  look = NULL,
  mod = NULL,
  name = "."
)
}
\arguments{
\item{pj}{an object of class \code{'pj'}.}

\item{x1}{numeric vector giving the horizontal start point of each segment.}

\item{y1}{numeric vector giving the vertical start point of each segment.}

\item{x2}{\code{NA} or a numeric vector giving the horizontal end point of
each segment.}

\item{y2}{\code{NA} or a numeric vectors giving the vertical end point of
each segment.}

\item{s}{\code{NA} or a numeric vector giving the slope of each segment.}

\item{a}{\code{NA} or a numeric vector giving the angle of each segment.}

\item{au}{\code{NA} or a character vector giving the angle units where
\code{'d'} indicates degrees, \code{'g'} indicates gradians, \code{'p'}
indicates proportion of a full rotation, and \code{'r'} indicates radians.}

\item{region}{either character scalar \code{"."} to indicate the most
recently used region, a positive integer to indicate the ID number of a
region (\code{0} indicates the background region), or a character scalar to
indicate the name of a region (\code{'bg'} indicates the background
region).}

\item{look}{an optional list containing up to four lists specifying looks to
apply to the shape. See the \emph{using \code{look} to make shapes appear}
section.}

\item{mod}{an optional list containing up to four lists specifying
modifications to apply to the shape. See the \emph{using \code{mod} to
transform shapes} section.}

\item{name}{character scalar indicating a name for the shape. The special
value \code{"."} indicates that the pj package create a default name for
the shape.}

\item{m}{\code{NA} or a numeric vector giving the magnitude (length) of each
segment.}
}
\value{
\code{pj} with the addition of the specified line segment shape(s).
}
\description{
A segments shape is a collection of lines between two points
(each) that use the same subset of locating arguments (\code{x}, \code{y},
\code{x2}, \code{y2}, \code{m}, \code{s}, \code{a}, and \code{au}).
}
\section{Using \code{look} to make shapes appear}{
 By itself, this function
simply defines the location of x-y pairs in the assigned region for the
specified shape. Various looks can be added to the shape using the
\code{look} parameter or via the \code{\link{add_ends}},
\code{\link{add_glyphs}}, \code{\link{add_labels}},
\code{\link{add_lines}}, and \code{\link{add_marks}} functions. Sub-lists
of the look argument must be named \code{'end'}, \code{'glyph'},
\code{'label'}, \code{'line'}, and/or \code{'mark'} to indicate that line
end marks, fill colors, glyphs (single characters), labels, lines, and/or
point marks be added to the shape.
}

\section{Argument recycling}{
 The argument set \code{\{x1, y1, x2, y2, m, s,
  a, au\}} is recycled.
}

\section{Using \code{mod} to transform shapes}{
 Shapes can be transformed in
four ways: dilation, reflection, rotation, and translation/. using the
\code{mod} argument or the \code{\link{mod_dilate}},
\code{\link{mod_reflect}}, \code{\link{mod_rotate}}, and
\code{\link{mod_translate}}. Sub-lists of the \code{mod} argument must be
named \code{'dilate'}, \code{'reflect'}, \code{'rotate'}, and/or
\code{'translate'} to indicate that a dilation, reflection, rotation, and
or translation be applied to the shape.
}

\seealso{
Other new_simples: 
\code{\link{new_points}()},
\code{\link{new_reflines}()}
}
\concept{new_simples}
