% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pj.R
\docType{data}
\name{pj.}
\alias{pj.}
\title{List of objects needed by the pjlot packages}
\format{
An object of class \code{list} of length 3.
}
\usage{
pj.
}
\description{
\code{pj.} contains the following elements:
\describe{
\item{blank}{blank pj plot template object.}
\item{patterns}{list of tibbles of patterns of argument characteristics.}
\item{vals}{list of character vectors of valid argumemt values.}
}
}
\keyword{datasets}
